Ref: move external services to their own file
