+++ packages/bitcore-wallet-service/src/lib/server.ts
@@ -1,5 +1,4 @@
 import * as async from 'async';
-import * as crypto from 'crypto'
 import {
   Constants as ConstantsCWC,
   Validation
@@ -11,6 +10,16 @@ import config from '../config';
 import logger from './logger';
 
 import { serverMessages as deprecatedServerMessage } from '../deprecated-serverMessages';
+import { BanxaService } from '../externalServices/banxa';
+import { ChangellyService } from '../externalServices/changelly';
+import { MoonpayService } from '../externalServices/moonpay';
+import { OneInchService } from '../externalServices/oneInch';
+import { RampService } from '../externalServices/ramp';
+import { SardineService } from '../externalServices/sardine';
+import { SimplexService } from '../externalServices/simplex';
+import { ThorswapService } from '../externalServices/thorswap';
+import { TransakService } from '../externalServices/transak';
+import { WyreService } from '../externalServices/wyre';
 import { serverMessages } from '../serverMessages';
 import { BCHAddressTranslator } from './bchaddresstranslator';
 import { BlockChainExplorer } from './blockchainexplorer';
@@ -129,6 +138,18 @@ export class WalletService implements IWalletService {
   copayerIsSupportStaff: boolean;
   copayerIsMarketingStaff: boolean;
   request: any;
+  externalServices: {
+    banxa: BanxaService;
+    changelly: ChangellyService;
+    moonpay: MoonpayService;
+    oneInch: OneInchService;
+    ramp: RampService;
+    sardine: SardineService;
+    simplex: SimplexService;
+    thorswap: ThorswapService;
+    transak: TransakService;
+    wyre: WyreService;
+  }
 
   constructor() {
     if (!initialized) {
@@ -145,6 +166,19 @@ export class WalletService implements IWalletService {
     // for testing
     //
     this.request = request;
+
+    this.externalServices = {
+      banxa: new BanxaService(),
+      changelly: new ChangellyService(),
+      moonpay: new MoonpayService(),
+      oneInch: new OneInchService(this.storage),
+      ramp: new RampService(),
+      sardine: new SardineService(),
+      simplex: new SimplexService(),
+      thorswap: new ThorswapService(),
+      transak: new TransakService(),
+      wyre: new WyreService(),
+    }
   }
   /**
    * Gets the current version of BWS
@@ -4841,2471 +4875,6 @@ export class WalletService implements IWalletService {
     return cb(null, externalServicesConfig);
   }
 
-  private banxaGetKeys(req) {
-    if (!config.banxa) throw new Error('Banxa missing credentials');
-
-    let env: 'sandbox' | 'production' | 'sandboxWeb' | 'productionWeb';
-    env = req.body.env === 'production' ? 'production' : 'sandbox';
-    if (req.body.context === 'web') {
-      env += 'Web';
-    }
-
-    delete req.body.env;
-    delete req.body.context;
-
-    const keys: {
-      API: string;
-      API_KEY: string;
-      SECRET_KEY: string;
-    } = {
-      API: config.banxa[env].api,
-      API_KEY: config.banxa[env].apiKey,
-      SECRET_KEY: config.banxa[env].secretKey
-    };
-
-    return keys;
-  }
-
-  private getBanxaSignature(method: 'get' | 'post', endpoint: string, apiKey: string, secret: string, body?: string) {
-    let signature, auth: string
-    const nonce = Date.now().toString();
-
-    switch (method) {
-      case 'get':
-        signature = 'GET' + '\n' + `/api${endpoint}` + '\n' + nonce;
-        break;
-      case 'post':
-        const stringifiedBody = body ? JSON.stringify(_.cloneDeep(body)) : '';
-        signature = 'POST' + '\n' + `/api${endpoint}` + '\n' + nonce + '\n' + stringifiedBody;
-        break;
-      default:
-        signature = undefined;
-        break;
-    }
-
-    const localSignature = crypto.createHmac('sha256', secret).update(signature).digest('hex');
-    auth = `${apiKey}:${localSignature}:${nonce}`
-    return auth;
-  }
-
-  banxaGetPaymentMethods(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.banxaGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      let qs = [];
-      if (req.body.source) qs.push('source=' + req.body.source);
-      if (req.body.target) qs.push('target=' + req.body.target);
-
-      const UriPath = `/payment-methods${qs.length > 0 ? '?' + qs.join('&') : ''}`;
-      const URL: string = API + UriPath;
-      const auth = this.getBanxaSignature('get', UriPath, API_KEY, SECRET_KEY);
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${auth}`
-      };
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  banxaGetQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.banxaGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      if (!checkRequired(req.body, ['source', 'target'])) {
-        return reject(new ClientError("Banxa's request missing arguments"));
-      }
-
-      let qs = [];
-      qs.push('source=' + req.body.source);
-      qs.push('target=' + req.body.target);
-
-      if (req.body.source_amount) qs.push('source_amount=' + req.body.source_amount);
-      if (req.body.target_amount) qs.push('target_amount=' + req.body.target_amount);
-      if (req.body.payment_method_id) qs.push('payment_method_id=' + req.body.payment_method_id);
-      if (req.body.account_reference) qs.push('account_reference=' + req.body.account_reference);
-      if (req.body.blockchain) qs.push('blockchain=' + req.body.blockchain);
-
-      const UriPath = `/prices${qs.length > 0 ? '?' + qs.join('&') : ''}`;
-      const URL: string = API + UriPath;
-      const auth = this.getBanxaSignature('get', UriPath, API_KEY, SECRET_KEY);
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${auth}`
-      };
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  banxaCreateOrder(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.banxaGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      if (!checkRequired(req.body, ['account_reference', 'source', 'target', 'wallet_address', 'return_url_on_success'])) {
-        return reject(new ClientError("Banxa's request missing arguments"));
-      }
-
-      delete req.body.payment_method_id;
-
-      const UriPath = '/orders';
-      const URL: string = API + UriPath;
-      const auth = this.getBanxaSignature('post', UriPath, API_KEY, SECRET_KEY, req.body);
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${auth}`
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  banxaGetOrder(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.banxaGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      if (!checkRequired(req.body, ['order_id'])) {
-        return reject(new ClientError("Banxa's request missing arguments"));
-      }
-
-      let qs = [];
-      if (req.body.fx_currency) qs.push('fx_currency=' + req.body.fx_currency);
-
-      const UriPath = `/orders/${req.body.order_id}${qs.length > 0 ? '?' + qs.join('&') : ''}`;
-      const URL: string = API + UriPath;
-      const auth = this.getBanxaSignature('get', UriPath, API_KEY, SECRET_KEY);
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${auth}`
-      };
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  private moonpayGetKeys(req) {
-    if (!config.moonpay) throw new Error('Moonpay missing credentials');
-
-    let env: 'sandbox' | 'production' | 'sandboxWeb' | 'productionWeb';
-    env = req.body.env === 'production' ? 'production' : 'sandbox';
-    if (req.body.context === 'web') {
-      env += 'Web';
-    }
-
-    delete req.body.env;
-    delete req.body.context;
-
-    const keys: {
-      API: string;
-      WIDGET_API: string;
-      SELL_WIDGET_API: string;
-      API_KEY: string;
-      SECRET_KEY: string;
-    } = {
-      API: config.moonpay[env].api,
-      WIDGET_API: config.moonpay[env].widgetApi,
-      SELL_WIDGET_API: config.moonpay[env].sellWidgetApi,
-      API_KEY: config.moonpay[env].apiKey,
-      SECRET_KEY: config.moonpay[env].secretKey
-    };
-
-    return keys;
-  }
-
-  moonpayGetCurrencies(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.moonpayGetKeys(req);
-      const API = keys.API;
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      const URL = API + '/v3/currencies/'
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  moonpayGetQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.moonpayGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      if (!checkRequired(req.body, ['currencyAbbreviation', 'baseCurrencyAmount', 'baseCurrencyCode'])) {
-        return reject(new ClientError("Moonpay's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      let qs = [];
-      qs.push('apiKey=' + API_KEY);
-      qs.push('baseCurrencyAmount=' + req.body.baseCurrencyAmount);
-      qs.push('baseCurrencyCode=' + req.body.baseCurrencyCode);
-
-      if (req.body.extraFeePercentage) qs.push('extraFeePercentage=' + req.body.extraFeePercentage);
-      if (req.body.paymentMethod) qs.push('paymentMethod=' + req.body.paymentMethod);
-      if (req.body.areFeesIncluded) qs.push('areFeesIncluded=' + req.body.areFeesIncluded);
-
-      const URL: string = API + `/v3/currencies/${req.body.currencyAbbreviation}/buy_quote/?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  moonpayGetSellQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.moonpayGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      if (!checkRequired(req.body, ['currencyAbbreviation', 'quoteCurrencyCode', 'baseCurrencyAmount'])) {
-        return reject(new ClientError("Moonpay's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      let qs = [];
-      qs.push('apiKey=' + API_KEY);
-      qs.push('quoteCurrencyCode=' + req.body.quoteCurrencyCode);
-      qs.push('baseCurrencyAmount=' + req.body.baseCurrencyAmount);
-
-      if (req.body.extraFeePercentage) qs.push('extraFeePercentage=' + req.body.extraFeePercentage);
-      if (req.body.payoutMethod) qs.push('payoutMethod=' + req.body.payoutMethod);
-
-      const URL: string = API + `/v3/currencies/${req.body.currencyAbbreviation}/sell_quote?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  moonpayGetCurrencyLimits(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.moonpayGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      if (!checkRequired(req.body, ['currencyAbbreviation', 'baseCurrencyCode'])) {
-        return reject(new ClientError("Moonpay's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      let qs = [];
-      qs.push('apiKey=' + API_KEY);
-      qs.push('baseCurrencyCode=' + encodeURIComponent(req.body.baseCurrencyCode));
-      if (req.body.areFeesIncluded) qs.push('areFeesIncluded=' + encodeURIComponent(req.body.areFeesIncluded));
-      if (req.body.paymentMethod) qs.push('paymentMethod=' + encodeURIComponent(req.body.paymentMethod));
-
-      const URL = API + `/v3/currencies/${req.body.currencyAbbreviation}/limits/?${qs.join('&')}`
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  moonpayGetSignedPaymentUrl(req): { urlWithSignature: string } {
-    const keys = this.moonpayGetKeys(req);
-    const SECRET_KEY = keys.SECRET_KEY;
-    const API_KEY = keys.API_KEY;
-    const WIDGET_API = keys.WIDGET_API;
-
-    if (
-      !checkRequired(req.body, [
-        'currencyCode',
-        'walletAddress',
-        'baseCurrencyCode',
-        'baseCurrencyAmount',
-        'externalTransactionId',
-        'redirectURL'
-      ])
-    ) {
-      throw new ClientError("Moonpay's request missing arguments");
-    }
-
-    const headers = {
-      'Content-Type': 'application/json'
-    };
-
-    let qs = [];
-    qs.push('apiKey=' + API_KEY);
-    qs.push('currencyCode=' + encodeURIComponent(req.body.currencyCode));
-    qs.push('walletAddress=' + encodeURIComponent(req.body.walletAddress));
-    qs.push('baseCurrencyCode=' + encodeURIComponent(req.body.baseCurrencyCode));
-    qs.push('baseCurrencyAmount=' + encodeURIComponent(req.body.baseCurrencyAmount));
-    qs.push('externalTransactionId=' + encodeURIComponent(req.body.externalTransactionId));
-    qs.push('redirectURL=' + encodeURIComponent(req.body.redirectURL));
-    if (req.body.lockAmount) qs.push('lockAmount=' + encodeURIComponent(req.body.lockAmount));
-    if (req.body.showWalletAddressForm)
-      qs.push('showWalletAddressForm=' + encodeURIComponent(req.body.showWalletAddressForm));
-    if (req.body.paymentMethod) qs.push('paymentMethod=' + encodeURIComponent(req.body.paymentMethod));
-    if (req.body.areFeesIncluded) qs.push('areFeesIncluded=' + encodeURIComponent(req.body.areFeesIncluded));
-
-    const URL_SEARCH: string = `?${qs.join('&')}`;
-
-    const URLSignatureHash: string = Bitcore.crypto.Hash.sha256hmac(
-      Buffer.from(URL_SEARCH),
-      Buffer.from(SECRET_KEY)
-    ).toString('base64');
-
-    const urlWithSignature = `${WIDGET_API}${URL_SEARCH}&signature=${encodeURIComponent(URLSignatureHash)}`;
-
-    return { urlWithSignature };
-  }
-
-  moonpayGetSellSignedPaymentUrl(req): { urlWithSignature: string } {
-    const keys = this.moonpayGetKeys(req);
-    const SECRET_KEY = keys.SECRET_KEY;
-    const API_KEY = keys.API_KEY;
-    const SELL_WIDGET_API = keys.SELL_WIDGET_API;
-
-    if (
-      !checkRequired(req.body, [
-        'baseCurrencyCode',
-        'baseCurrencyAmount',
-        'externalTransactionId',
-        'redirectURL',
-      ])
-    ) {
-      throw new ClientError("Moonpay's request missing arguments");
-    }
-
-    const headers = {
-      'Content-Type': 'application/json'
-    };
-
-    let qs = [];
-    qs.push('apiKey=' + API_KEY);
-    qs.push('baseCurrencyCode=' + encodeURIComponent(req.body.baseCurrencyCode));
-    qs.push('baseCurrencyAmount=' + encodeURIComponent(req.body.baseCurrencyAmount));
-    qs.push('externalTransactionId=' + encodeURIComponent(req.body.externalTransactionId));
-    qs.push('redirectURL=' + encodeURIComponent(req.body.redirectURL));
-
-    if (req.body.quoteCurrencyCode) qs.push('quoteCurrencyCode=' + encodeURIComponent(req.body.quoteCurrencyCode));
-    if (req.body.paymentMethod) qs.push('paymentMethod=' + encodeURIComponent(req.body.paymentMethod));
-    if (req.body.externalCustomerId) qs.push('externalCustomerId=' + encodeURIComponent(req.body.externalCustomerId));
-    if (req.body.refundWalletAddress) qs.push('refundWalletAddress=' + encodeURIComponent(req.body.refundWalletAddress));
-    if (req.body.lockAmount) qs.push('lockAmount=' + encodeURIComponent(req.body.lockAmount));
-    if (req.body.colorCode) qs.push('colorCode=' + encodeURIComponent(req.body.colorCode));
-    if (req.body.theme) qs.push('theme=' + encodeURIComponent(req.body.theme));
-    if (req.body.language) qs.push('language=' + encodeURIComponent(req.body.language));
-    if (req.body.email) qs.push('email=' + encodeURIComponent(req.body.email));
-    if (req.body.externalCustomerId) qs.push('externalCustomerId=' + encodeURIComponent(req.body.externalCustomerId));
-    if (req.body.showWalletAddressForm)
-      qs.push('showWalletAddressForm=' + encodeURIComponent(req.body.showWalletAddressForm));
-    if (req.body.unsupportedRegionRedirectUrl) qs.push('unsupportedRegionRedirectUrl=' + encodeURIComponent(req.body.unsupportedRegionRedirectUrl));
-    if (req.body.skipUnsupportedRegionScreen) qs.push('skipUnsupportedRegionScreen=' + encodeURIComponent(req.body.skipUnsupportedRegionScreen));
-
-    const URL_SEARCH: string = `?${qs.join('&')}`;
-
-    const URLSignatureHash: string = Bitcore.crypto.Hash.sha256hmac(
-      Buffer.from(URL_SEARCH),
-      Buffer.from(SECRET_KEY)
-    ).toString('base64');
-
-    const urlWithSignature = `${SELL_WIDGET_API}${URL_SEARCH}&signature=${encodeURIComponent(URLSignatureHash)}`;
-
-    return { urlWithSignature };
-  }
-
-  moonpayGetTransactionDetails(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.moonpayGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      if (!checkRequired(req.body, ['transactionId']) && !checkRequired(req.body, ['externalId'])) {
-        return reject(new ClientError("Moonpay's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-      let URL: string;
-
-      let qs = [];
-      qs.push('apiKey=' + API_KEY);
-      if (req.body.transactionId) {
-        URL = API + `/v1/transactions/${req.body.transactionId}?${qs.join('&')}`;
-      } else if (req.body.externalId) {
-        URL = API + `/v1/transactions/ext/${req.body.externalId}?${qs.join('&')}`;
-      }
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  moonpayGetSellTransactionDetails(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.moonpayGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      if (!checkRequired(req.body, ['transactionId']) && !checkRequired(req.body, ['externalId'])) {
-        return reject(new ClientError("Moonpay's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-      let URL: string;
-
-      let qs = [];
-      qs.push('apiKey=' + API_KEY);
-      if (req.body.transactionId) {
-        URL = API + `/v3/sell_transactions/${req.body.transactionId}?${qs.join('&')}`;
-      } else if (req.body.externalId) {
-        URL = API + `/v3/sell_transactions/ext/${req.body.externalId}?${qs.join('&')}`;
-      }
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  moonpayCancelSellTransaction(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.moonpayGetKeys(req);
-      const API = keys.API;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      if (!checkRequired(req.body, ['transactionId']) && !checkRequired(req.body, ['externalId'])) {
-        return reject(new ClientError("Moonpay's request missing arguments"));
-      }
-
-      const headers = {
-        Authorization: 'Api-Key ' + SECRET_KEY,
-        Accept: 'application/json'
-      };
-      let URL: string;
-
-      if (req.body.transactionId) {
-        URL = API + `/v3/sell_transactions/${req.body.transactionId}`;
-      } else if (req.body.externalId) {
-        URL = API + `/v3/sell_transactions/ext/${req.body.externalId}`;
-      }
-
-      this.request.delete(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  moonpayGetAccountDetails(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.moonpayGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      let qs = [];
-      qs.push('apiKey=' + API_KEY);
-
-      const URL = API + `/v3/accounts/me?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  private rampGetKeys(req) {
-    if (!config.ramp) throw new Error('Ramp missing credentials');
-
-    let env: 'sandbox' | 'production' | 'sandboxWeb' | 'productionWeb';
-    env = req.body.env === 'production' ? 'production' : 'sandbox';
-    if (req.body.context === 'web') {
-      env += 'Web';
-    }
-    delete req.body.env;
-    delete req.body.context;
-
-    const keys: {
-      API: string;
-      WIDGET_API: string;
-      API_KEY: string;
-    } = {
-      API: config.ramp[env].api,
-      WIDGET_API: config.ramp[env].widgetApi,
-      API_KEY: config.ramp[env].apiKey,
-    };
-
-    return keys;
-  }
-
-  rampGetQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.rampGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      if (!checkRequired(req.body, ['cryptoAssetSymbol', 'fiatValue', 'fiatCurrency'])) {
-        return reject(new ClientError("Ramp's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      const URL: string = API + `/host-api/v3/onramp/quote/all?hostApiKey=${API_KEY}`;
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  rampGetSellQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.rampGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      /*
-        * Although fiatValue and cryptoAmount are not both required, you need to pass one of them.
-        * cryptoAmount - should be passed in token wei - e.g. for 1ETH cryptoAmount: 1000000000000000000
-        * cryptoAmount?: string;
-        * fiatValue?: number;
-      */
-      if (!checkRequired(req.body, ['cryptoAssetSymbol', 'fiatCurrency']) || (!checkRequired(req.body, ['fiatValue']) && !checkRequired(req.body, ['cryptoAmount']))) {
-        return reject(new ClientError("Ramp's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      const URL: string = API + `/host-api/v3/offramp/quote/all?hostApiKey=${API_KEY}`;
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  rampGetSignedPaymentUrl(req): { urlWithSignature: string } {
-    const webRequiredParams = [
-      'selectedCountryCode',
-    ];
-    const appRequiredParams = [
-      'enabledFlows',
-      'defaultFlow',
-      'selectedCountryCode',
-      'defaultAsset',
-    ];
-    const extraRequiredParams = req.body.flow && req.body.flow === 'sell' ? ['offrampAsset'] : ['finalUrl', 'userAddress', 'swapAmount','swapAsset'];
-    appRequiredParams.concat(extraRequiredParams);
-
-    const requiredParams = req.body.context === 'web' ? webRequiredParams : appRequiredParams;
-    const keys = this.rampGetKeys(req);
-    const API_KEY = keys.API_KEY;
-    const WIDGET_API = keys.WIDGET_API;
-
-    if (
-      !checkRequired(req.body, requiredParams)
-    ) {
-      throw new ClientError("Ramp's request missing arguments");
-    }
-
-    const headers = {
-      'Content-Type': 'application/json'
-    };
-
-    let qs = [];
-    qs.push('hostApiKey=' + API_KEY);
-    qs.push('selectedCountryCode=' + encodeURIComponent(req.body.selectedCountryCode));
-    if (req.body.finalUrl) qs.push('finalUrl=' + encodeURIComponent(req.body.finalUrl));
-    if (req.body.userAddress) qs.push('userAddress=' + encodeURIComponent(req.body.userAddress));
-    if (req.body.swapAsset) qs.push('swapAsset=' + encodeURIComponent(req.body.swapAsset));
-    if (req.body.offrampAsset) qs.push('offrampAsset=' + encodeURIComponent(req.body.offrampAsset));
-    if (req.body.enabledFlows) qs.push('enabledFlows=' + encodeURIComponent(req.body.enabledFlows));
-    if (req.body.defaultFlow) qs.push('defaultFlow=' + encodeURIComponent(req.body.defaultFlow));
-    if (req.body.hostLogoUrl) qs.push('hostLogoUrl=' + encodeURIComponent(req.body.hostLogoUrl));
-    if (req.body.hostAppName) qs.push('hostAppName=' + encodeURIComponent(req.body.hostAppName));
-    if (req.body.swapAmount) qs.push('swapAmount=' + encodeURIComponent(req.body.swapAmount));
-    if (req.body.fiatValue) qs.push('fiatValue=' + encodeURIComponent(req.body.fiatValue));
-    if (req.body.fiatCurrency) qs.push('fiatCurrency=' + encodeURIComponent(req.body.fiatCurrency));
-    if (req.body.defaultAsset) qs.push('defaultAsset=' + encodeURIComponent(req.body.defaultAsset));
-    if (req.body.userEmailAddress) qs.push('userEmailAddress=' + encodeURIComponent(req.body.userEmailAddress));
-    if (req.body.useSendCryptoCallback) qs.push('useSendCryptoCallback=' + encodeURIComponent(req.body.useSendCryptoCallback));
-    if (req.body.paymentMethodType) qs.push('paymentMethodType=' + encodeURIComponent(req.body.paymentMethodType));
-    if (req.body.hideExitButton) qs.push('hideExitButton=' + encodeURIComponent(req.body.hideExitButton));
-    if (req.body.variant) qs.push('variant=' + encodeURIComponent(req.body.variant));
-    if (req.body.useSendCryptoCallbackVersion) qs.push('useSendCryptoCallbackVersion=' + encodeURIComponent(req.body.useSendCryptoCallbackVersion));
-
-    const URL_SEARCH: string = `?${qs.join('&')}`;
-
-    const urlWithSignature = `${WIDGET_API}${URL_SEARCH}`;
-
-    return { urlWithSignature };
-  }
-
-  rampGetAssets(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.rampGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-      let URL: string;
-
-      let qs = [];
-      // "Buy" and "Sell" features use the same properties. Use "flow" to target the correct endpoint
-      qs.push('hostApiKey=' + API_KEY);
-      if (req.body.currencyCode) qs.push('currencyCode=' + encodeURIComponent(req.body.currencyCode));
-      if (req.body.withDisabled) qs.push('withDisabled=' + encodeURIComponent(req.body.withDisabled));
-      if (req.body.withHidden) qs.push('withHidden=' + encodeURIComponent(req.body.withHidden));
-      if (req.body.useIp) {
-        const ip = Utils.getIpFromReq(req);
-        qs.push('userIp=' + encodeURIComponent(ip));
-      }
-
-      URL = API + `/host-api/v3${req.body.flow && req.body.flow === 'sell' ? '/offramp' : ''}/assets?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  rampGetSellTransactionDetails(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.rampGetKeys(req);
-      const API = keys.API;
-
-      if (!checkRequired(req.body, ['id', 'saleViewToken'])) {
-        return reject(new ClientError("Ramp's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-      let URL: string;
-
-      let qs = [];
-      qs.push('secret=' + req.body.saleViewToken);
-
-      URL = API + `/host-api/v3/offramp/sale/${req.body.id}?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  private sardineGetKeys(req) {
-    if (!config.sardine) throw new Error('Sardine missing credentials');
-
-    let env: 'sandbox' | 'production' | 'sandboxWeb' | 'productionWeb';
-    env = req.body.env === 'production' ? 'production' : 'sandbox';
-    if (req.body.context === 'web') {
-      env += 'Web';
-    }
-    delete req.body.env;
-    delete req.body.context;
-
-    const keys: {
-      API: string;
-      SECRET_KEY: string;
-      CLIENT_ID: string;
-    } = {
-      API: config.sardine[env].api,
-      SECRET_KEY: config.sardine[env].secretKey,
-      CLIENT_ID: config.sardine[env].clientId,
-    };
-
-    return keys;
-  }
-
-  sardineGetQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.sardineGetKeys(req);
-      const API = keys.API;
-      const CLIENT_ID = keys.CLIENT_ID;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      if (!checkRequired(req.body, ['asset_type', 'network', 'total'])) {
-        return reject(new ClientError("Sardine's request missing arguments"));
-      }
-
-      const secret = `${CLIENT_ID}:${SECRET_KEY}`;
-      const secretBase64 = Buffer.from(secret).toString('base64');
-
-      const headers = {
-        Accept: 'application/json',
-        Authorization: `Basic ${secretBase64}`,
-      };
-
-      let qs = [];
-      qs.push('asset_type=' + req.body.asset_type);
-      qs.push('network=' + req.body.network);
-      qs.push('total=' + req.body.total);
-
-      if (req.body.currency) qs.push('currency=' + req.body.currency);
-      if (req.body.paymentType) qs.push('paymentType=' + req.body.paymentType);
-      if (req.body.quote_type) qs.push('quote_type=' + req.body.quote_type);
-
-      const URL: string = API + `/v1/quotes?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  sardineGetCurrencyLimits(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.sardineGetKeys(req);
-      const API = keys.API;
-      const CLIENT_ID = keys.CLIENT_ID;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      const secret = `${CLIENT_ID}:${SECRET_KEY}`;
-      const secretBase64 = Buffer.from(secret).toString('base64');
-
-      const headers = {
-        Accept: 'application/json',
-        Authorization: `Basic ${secretBase64}`,
-      };
-
-      const URL: string = API + '/v1/fiat-currencies';
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  sardineGetToken(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.sardineGetKeys(req);
-      const API = keys.API;
-      const CLIENT_ID = keys.CLIENT_ID;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      if (!checkRequired(req.body, ['referenceId', 'externalUserId', 'customerId'])) {
-        return reject(new ClientError("Sardine's request missing arguments"));
-      }
-
-      const secret = `${CLIENT_ID}:${SECRET_KEY}`;
-      const secretBase64 = Buffer.from(secret).toString('base64');
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'Authorization': `Basic ${secretBase64}`,
-      };
-
-      const URL: string = API + '/v1/auth/client-tokens';
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  sardineGetSupportedTokens(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.sardineGetKeys(req);
-      const API = keys.API;
-      const CLIENT_ID = keys.CLIENT_ID;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      const secret = `${CLIENT_ID}:${SECRET_KEY}`;
-      const secretBase64 = Buffer.from(secret).toString('base64');
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'Authorization': `Basic ${secretBase64}`,
-      };
-
-      const URL: string = API + '/v1/supported-tokens';
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  sardineGetOrdersDetails(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.sardineGetKeys(req);
-      const API = keys.API;
-      const CLIENT_ID = keys.CLIENT_ID;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      if (!checkRequired(req.body, ['orderId']) && !checkRequired(req.body, ['externalUserId']) && !checkRequired(req.body, ['referenceId'])) {
-        return reject(new ClientError("Sardine's request missing arguments"));
-      }
-
-      const secret = `${CLIENT_ID}:${SECRET_KEY}`;
-      const secretBase64 = Buffer.from(secret).toString('base64');
-
-      const headers = {
-        Accept: 'application/json',
-        Authorization: `Basic ${secretBase64}`,
-      };
-
-      let qs = [];
-      let URL: string;
-
-      if (req.body.orderId) {
-        URL = API + `/v1/orders/${req.body.orderId}`;
-      } else if (req.body.externalUserId || req.body.referenceId) {
-        if (req.body.externalUserId) qs.push('externalUserId=' + req.body.externalUserId);
-        if (req.body.referenceId) qs.push('referenceId=' + req.body.referenceId);
-        if (req.body.startDate) qs.push('startDate=' + req.body.startDate);
-        if (req.body.endDate) qs.push('endDate=' + req.body.endDate);
-        if (req.body.limit) qs.push('limit=' + req.body.limit);
-
-        URL = API + `/v1/orders?${qs.join('&')}`;
-      }
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  private simplexGetKeys(req) {
-    if (!config.simplex) throw new Error('Simplex missing credentials');
-
-    let env: 'sandbox' | 'production' | 'sandboxWeb' | 'productionWeb';
-    env = req.body.env === 'production' ? 'production' : 'sandbox';
-    if (req.body.context === 'web') {
-      env += 'Web';
-    }
-
-    delete req.body.env;
-    delete req.body.context;
-
-    const keys = {
-      API: config.simplex[env].api,
-      API_SELL: config.simplex[env].apiSell,
-      API_KEY: config.simplex[env].apiKey,
-      PUBLIC_KEY: config.simplex[env].publicKey,
-      APP_PROVIDER_ID: config.simplex[env].appProviderId,
-      APP_SELL_REF_ID: config.simplex[env].appSellRefId
-    };
-
-    return keys;
-  }
-
-  simplexGetCurrencies(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.simplexGetKeys(req);
-      const API = keys.API;
-      const PUBLIC_KEY = keys.PUBLIC_KEY;
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      const URL = API + `/v2/supported_crypto_currencies?public_key=${PUBLIC_KEY}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  simplexGetQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.simplexGetKeys(req);
-
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-      const ip = Utils.getIpFromReq(req);
-
-      req.body.client_ip = ip;
-      req.body.wallet_id = keys.APP_PROVIDER_ID;
-
-      const headers = {
-        'Content-Type': 'application/json',
-        Authorization: 'ApiKey ' + API_KEY
-      };
-
-      if (req.body && req.body.payment_methods && Array.isArray(req.body.payment_methods)) {
-        // Workaround to fix older versions of the app
-        req.body.payment_methods = req.body.payment_methods.map(item => item === 'simplex_account' ? 'sepa_open_banking' : item);
-      }
-
-      this.request.post(
-        API + '/wallet/merchant/v2/quote',
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : null);
-          }
-        }
-      );
-    });
-  }
-
-  simplexGetSellQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.simplexGetKeys(req);
-
-      const API = keys.API_SELL;
-      const API_KEY = keys.API_KEY;
-
-      if (!checkRequired(req.body, ['base_currency', 'base_amount', 'quote_currency', 'pp_payment_method'])) {
-        return reject(new ClientError("Simplex's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json',
-        Authorization: 'ApiKey ' + API_KEY,
-      };
-
-      if (req.body.userCountry && typeof req.body.userCountry === 'string') {
-        headers['x-country-code'] = req.body.userCountry.toUpperCase();
-      }
-
-      let qs = [];
-      qs.push('base_currency=' + req.body.base_currency);
-      qs.push('base_amount=' + req.body.base_amount);
-      qs.push('quote_currency=' + req.body.quote_currency);
-      qs.push('pp_payment_method=' + req.body.pp_payment_method);
-
-      const URL: string = API + `/v3/quote?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  simplexPaymentRequest(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.simplexGetKeys(req);
-
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-      const appProviderId = keys.APP_PROVIDER_ID;
-      const paymentId = Uuid.v4();
-      const orderId = Uuid.v4();
-      const apiHost = keys.API;
-      const ip = Utils.getIpFromReq(req);
-
-      if (
-        !checkRequired(req.body, ['account_details', 'transaction_details']) &&
-        !checkRequired(req.body.transaction_details, ['payment_details'])
-      ) {
-        return reject(new ClientError("Simplex's request missing arguments"));
-      }
-
-      req.body.account_details.app_provider_id = appProviderId;
-      req.body.account_details.signup_login = {
-        ip,
-        location: '',
-        uaid: '',
-        accept_language: 'de,en-US;q=0.7,en;q=0.3',
-        http_accept_language: 'de,en-US;q=0.7,en;q=0.3',
-        user_agent: req.body.account_details.signup_login ? req.body.account_details.signup_login.user_agent : '', // Format: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:67.0) Gecko/20100101 Firefox/67.0'
-        cookie_session_id: '',
-        timestamp: req.body.account_details.signup_login ? req.body.account_details.signup_login.timestamp : ''
-      };
-
-      req.body.transaction_details.payment_details.payment_id = paymentId;
-      req.body.transaction_details.payment_details.order_id = orderId;
-
-      const headers = {
-        'Content-Type': 'application/json',
-        Authorization: 'ApiKey ' + API_KEY
-      };
-
-      this.request.post(
-        API + '/wallet/merchant/v2/payments/partner/data',
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            data.body.payment_id = paymentId;
-            data.body.order_id = orderId;
-            data.body.app_provider_id = appProviderId;
-            data.body.api_host = apiHost;
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  simplexSellPaymentRequest(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.simplexGetKeys(req);
-
-      const API = keys.API_SELL;
-      const API_KEY = keys.API_KEY;
-      const appSellRefId = keys.APP_SELL_REF_ID;
-
-      if (
-        !checkRequired(req.body, ['referer_url', 'return_url']) ||
-        !checkRequired(req.body.txn_details, ['quote_id'])
-      ) {
-        return reject(new ClientError("Simplex's request missing arguments"));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json',
-        Authorization: 'ApiKey ' + API_KEY,
-      };
-
-      if (req.body.userCountry && typeof req.body.userCountry === 'string') {
-        headers['x-country-code'] = req.body.userCountry.toUpperCase();
-      }
-
-      this.request.post(
-        API + '/v3/initiate-sell/widget',
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            data.body.app_sell_ref_id = appSellRefId;
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  simplexGetEvents(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      if (!config.simplex) return reject(new Error('Simplex missing credentials'));
-      if (!req.env || (req.env != 'sandbox' && req.env != 'production'))
-        return reject(new Error("Simplex's request wrong environment"));
-
-      const API = config.simplex[req.env].api;
-      const API_KEY = config.simplex[req.env].apiKey;
-      const headers = {
-        'Content-Type': 'application/json',
-        Authorization: 'ApiKey ' + API_KEY
-      };
-
-      this.request.get(
-        API + '/wallet/merchant/v2/events',
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : null);
-          } else {
-            return resolve(data.body ? data.body : null);
-          }
-        }
-      );
-    });
-  }
-
-  private wyreGetKeys(req) {
-    if (!config.wyre) throw new Error('Wyre missing credentials');
-
-    let env = 'sandbox';
-    if (req.body.env && req.body.env == 'production') {
-      env = 'production';
-    }
-    delete req.body.env;
-
-    const keys = {
-      API: config.wyre[env].api,
-      API_KEY: config.wyre[env].apiKey,
-      SECRET_API_KEY: config.wyre[env].secretApiKey,
-      ACCOUNT_ID: config.wyre[env].appProviderAccountId
-    };
-
-    return keys;
-  }
-
-  private thorswapGetKeys(req) {
-    if (!config.thorswap) throw new Error('Thorswap missing credentials');
-
-    let env: 'sandbox' | 'production' | 'sandboxWeb' | 'productionWeb';
-    env = req.body.env === 'production' ? 'production' : 'sandbox';
-    if (req.body.context === 'web') {
-      env += 'Web';
-    }
-
-    delete req.body.env;
-    delete req.body.context;
-
-    const keys: {
-      API: string;
-      API_KEY: string;
-      SECRET_KEY: string;
-      REFERER: string;
-    } = {
-      API: config.thorswap[env].api,
-      API_KEY: config.thorswap[env].apiKey,
-      SECRET_KEY: config.thorswap[env].secretKey,
-      REFERER: config.thorswap[env].referer
-    };
-
-    return keys;
-  }
-
-  thorswapGetSupportedChains(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.thorswapGetKeys(req);
-      const API = keys.API;
-      const REFERER = keys.REFERER;
-      const API_KEY = keys.API_KEY;
-
-      const headers = {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'Referer': REFERER,
-        'x-api-key': API_KEY
-      };
-
-      const uriPath: string = req?.body?.includeDetails ? '/tokenlist/utils/chains/details' : '/tokenlist/utils/chains';
-      const URL: string = API + uriPath;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  thorswapGetCryptoCurrencies(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.thorswapGetKeys(req);
-      const API = keys.API;
-      const REFERER = keys.REFERER;
-      const API_KEY = keys.API_KEY;
-
-      const headers = {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'Referer': REFERER,
-        'x-api-key': API_KEY
-      };
-
-      let qs = [];
-      qs.push('categories=' + (req?.body?.categories ?? 'all'));
-
-      const uriPath: string = req?.body?.includeDetails ? '/tokenlist/utils/currencies/details' : '/tokenlist/utils/currencies';
-      const URL: string = API + `${uriPath}?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  thorswapGetSwapQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.thorswapGetKeys(req);
-      const API = keys.API;
-      const REFERER = keys.REFERER;
-      const API_KEY = keys.API_KEY;
-
-      const headers = {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'Referer': REFERER,
-        'x-api-key': API_KEY
-      };
-
-      let qs = [];
-      if (!checkRequired(req.body, ['sellAsset', 'buyAsset', 'sellAmount'])) {
-        return reject(new ClientError("Thorswap's request missing arguments"));
-      }
-      qs.push('sellAsset=' + req.body.sellAsset);
-      qs.push('buyAsset=' + req.body.buyAsset);
-      qs.push('sellAmount=' + req.body.sellAmount);
-      if (req.body.senderAddress) qs.push('senderAddress=' + req.body.senderAddress);
-      if (req.body.recipientAddress) qs.push('recipientAddress=' + req.body.recipientAddress);
-      if (req.body.slippage) qs.push('slippage=' + req.body.slippage);
-      if (req.body.limit) qs.push('limit=' + req.body.limit);
-      if (req.body.providers) qs.push('providers=' + req.body.providers);
-      if (req.body.subProviders) qs.push('subProviders=' + req.body.subProviders);
-      if (req.body.preferredProvider) qs.push('preferredProvider=' + req.body.preferredProvider);
-      if (req.body.affiliateAddress) qs.push('affiliateAddress=' + req.body.affiliateAddress);
-      if (req.body.affiliateBasisPoints) qs.push('affiliateBasisPoints=' + req.body.affiliateBasisPoints);
-      if (req.body.isAffiliateFeeFlat) qs.push('isAffiliateFeeFlat=' + req.body.isAffiliateFeeFlat);
-      if (req.body.allowSmartContractRecipient) qs.push('allowSmartContractRecipient=' + req.body.allowSmartContractRecipient);
-
-      const URL: string = API + `/aggregator/tokens/quote?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  thorswapGetSwapTx(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.thorswapGetKeys(req);
-      const API = keys.API;
-      const REFERER = keys.REFERER;
-      const API_KEY = keys.API_KEY;
-
-      const headers = {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'Referer': REFERER,
-        'x-api-key': API_KEY
-      };
-
-      if (!checkRequired(req.body, ['hash']) && !checkRequired(req.body, ['txn'])) {
-        return reject(new ClientError("Thorswap's request missing arguments"));
-      }
-
-      this.request.post(
-        API + '/tracker/v2/txn',
-        // API + '/apiusage/v2/txn',
-        // 'https://api.swapkit.dev/track',
-        // /apiusage/v2/txn
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  private transakGetKeys(req) {
-    if (!config.transak) throw new Error('Transak missing credentials');
-
-    let env: 'sandbox' | 'production' | 'sandboxWeb' | 'productionWeb';
-    env = req.body.env === 'production' ? 'production' : 'sandbox';
-    if (req.body.context === 'web') {
-      env += 'Web';
-    }
-
-    delete req.body.env;
-    delete req.body.context;
-
-    const keys: {
-      API: string;
-      API_KEY: string;
-      SECRET_KEY: string;
-      WIDGET_API: string;
-    } = {
-      API: config.transak[env].api,
-      API_KEY: config.transak[env].apiKey,
-      SECRET_KEY: config.transak[env].secretKey,
-      WIDGET_API: config.transak[env].widgetApi
-    };
-
-    return keys;
-  }
-
-  transakGetAccessToken(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.transakGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-      const SECRET_KEY = keys.SECRET_KEY;
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'api-secret': SECRET_KEY,
-      };
-
-      req.body = {
-        apiKey: API_KEY
-      }
-
-      const URL: string = API + '/partners/api/v2/refresh-token';
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  transakGetCryptoCurrencies(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.transakGetKeys(req);
-      const API = keys.API;
-
-      const headers = {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-      };
-
-      const URL: string = API + '/api/v2/currencies/crypto-currencies';
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  transakGetFiatCurrencies(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.transakGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      const headers = {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-      };
-
-      const URL: string = API + `/api/v2/currencies/fiat-currencies?apiKey=${API_KEY}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  transakGetQuote(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.transakGetKeys(req);
-      const API = keys.API;
-      const API_KEY = keys.API_KEY;
-
-      if (!checkRequired(req.body, ['fiatCurrency', 'cryptoCurrency', 'network', 'paymentMethod'])) {
-        return reject(new ClientError("Transak's request missing arguments"));
-      }
-
-      const headers = {
-        Accept: 'application/json',
-      };
-
-      let qs = [];
-      qs.push('partnerApiKey=' + API_KEY);
-      qs.push('fiatCurrency=' + req.body.fiatCurrency);
-      qs.push('cryptoCurrency=' + req.body.cryptoCurrency);
-      qs.push('isBuyOrSell=BUY');
-      qs.push('network=' + req.body.network);
-      qs.push('paymentMethod=' + req.body.paymentMethod);
-
-      if (req.body.fiatAmount) qs.push('fiatAmount=' + req.body.fiatAmount);
-      if (req.body.cryptoAmount) qs.push('cryptoAmount=' + req.body.cryptoAmount);
-
-      const URL: string = API + `/api/v2/currencies/price?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-  transakGetSignedPaymentUrl(req): { urlWithSignature: string } {
-    const appRequiredParams = [
-      'walletAddress',
-      'redirectURL',
-      'fiatAmount',
-      'fiatCurrency',
-      'network',
-      'cryptoCurrencyCode',
-      'partnerOrderId',
-      'partnerCustomerId',
-    ];
-
-    const requiredParams = req.body.context === 'web' ? [] : appRequiredParams;
-    const keys = this.transakGetKeys(req);
-    const API_KEY = keys.API_KEY;
-    const WIDGET_API = keys.WIDGET_API;
-
-    if (
-      !checkRequired(req.body, requiredParams)
-    ) {
-      throw new ClientError("Transak's request missing arguments");
-    }
-
-    const headers = {
-      'Content-Type': 'application/json'
-    };
-
-    let qs = [];
-    // Recommended parameters to customize from the app
-    if (req.body.walletAddress) qs.push('walletAddress=' + encodeURIComponent(req.body.walletAddress));
-    if (req.body.disableWalletAddressForm) qs.push('disableWalletAddressForm=' + encodeURIComponent(req.body.disableWalletAddressForm));
-    if (req.body.redirectURL) qs.push('redirectURL=' + encodeURIComponent(req.body.redirectURL));
-    if (req.body.exchangeScreenTitle) qs.push('exchangeScreenTitle=' + encodeURIComponent(req.body.exchangeScreenTitle));
-    if (req.body.fiatAmount) qs.push('fiatAmount=' + encodeURIComponent(req.body.fiatAmount));
-    if (req.body.fiatCurrency) qs.push('fiatCurrency=' + encodeURIComponent(req.body.fiatCurrency));
-    if (req.body.network) qs.push('network=' + encodeURIComponent(req.body.network));
-    if (req.body.paymentMethod) qs.push('paymentMethod=' + encodeURIComponent(req.body.paymentMethod));
-    if (req.body.cryptoCurrencyCode) qs.push('cryptoCurrencyCode=' + encodeURIComponent(req.body.cryptoCurrencyCode));
-    if (req.body.cryptoCurrencyList) qs.push('cryptoCurrencyList=' + encodeURIComponent(req.body.cryptoCurrencyList));
-    if (req.body.hideExchangeScreen) qs.push('hideExchangeScreen=' + encodeURIComponent(req.body.hideExchangeScreen));
-    if (req.body.themeColor) qs.push('themeColor=' + encodeURIComponent(req.body.themeColor));
-    if (req.body.hideMenu) qs.push('hideMenu=' + encodeURIComponent(req.body.hideMenu));
-    if (req.body.partnerOrderId) qs.push('partnerOrderId=' + encodeURIComponent(req.body.partnerOrderId));
-    if (req.body.partnerCustomerId) qs.push('partnerCustomerId=' + encodeURIComponent(req.body.partnerCustomerId));
-    // Other parameters
-    if (req.body.environment) qs.push('environment=' + encodeURIComponent(req.body.environment));
-    if (req.body.widgetHeight) qs.push('widgetHeight=' + encodeURIComponent(req.body.widgetHeight));
-    if (req.body.widgetWidth) qs.push('widgetWidth=' + encodeURIComponent(req.body.widgetWidth));
-    if (req.body.productsAvailed) qs.push('productsAvailed=' + encodeURIComponent(req.body.productsAvailed));
-    if (req.body.defaultFiatAmount) qs.push('defaultFiatAmount=' + encodeURIComponent(req.body.defaultFiatAmount));
-    if (req.body.countryCode) qs.push('countryCode=' + encodeURIComponent(req.body.countryCode));
-    if (req.body.excludeFiatCurrencies) qs.push('excludeFiatCurrencies=' + encodeURIComponent(req.body.excludeFiatCurrencies));
-    if (req.body.defaultNetwork) qs.push('defaultNetwork=' + encodeURIComponent(req.body.defaultNetwork));
-    if (req.body.networks) qs.push('networks=' + encodeURIComponent(req.body.networks));
-    if (req.body.defaultPaymentMethod) qs.push('defaultPaymentMethod=' + encodeURIComponent(req.body.defaultPaymentMethod));
-    if (req.body.disablePaymentMethods) qs.push('disablePaymentMethods=' + encodeURIComponent(req.body.disablePaymentMethods));
-    if (req.body.defaultCryptoAmount) qs.push('defaultCryptoAmount=' + encodeURIComponent(req.body.defaultCryptoAmount));
-    if (req.body.cryptoAmount) qs.push('cryptoAmount=' + encodeURIComponent(req.body.cryptoAmount));
-    if (req.body.defaultCryptoCurrency) qs.push('defaultCryptoCurrency=' + encodeURIComponent(req.body.defaultCryptoCurrency));
-    if (req.body.isFeeCalculationHidden) qs.push('isFeeCalculationHidden=' + encodeURIComponent(req.body.isFeeCalculationHidden));
-    if (req.body.walletAddressesData) qs.push('walletAddressesData=' + encodeURIComponent(req.body.walletAddressesData));
-    if (req.body.email) qs.push('email=' + encodeURIComponent(req.body.email));
-    if (req.body.userData) qs.push('userData=' + encodeURIComponent(req.body.userData));
-    if (req.body.isAutoFillUserData) qs.push('isAutoFillUserData=' + encodeURIComponent(req.body.isAutoFillUserData));
-
-    const URL_SEARCH: string = `?apiKey=${API_KEY}&${qs.join('&')}`;
-
-    const urlWithSignature = `${WIDGET_API}${URL_SEARCH}`;
-
-    return { urlWithSignature };
-  }
-
-  transakGetOrderDetails(req): Promise<any> {
-    return new Promise(async (resolve, reject) => {
-      const env = _.cloneDeep(req.body.env);
-      const keys = this.transakGetKeys(req);
-      const API = keys.API;
-
-      if (!checkRequired(req.body, ['orderId'])) {
-        return reject(new ClientError("Transak's request missing arguments"));
-      }
-
-      let accessToken;
-      if (req.body.accessToken) {
-        accessToken = req.body.accessToken;
-      } else {
-        try {
-          const accessTokenData = await this.transakGetAccessToken({ body: env });
-          accessToken = accessTokenData?.data?.accessToken;
-        } catch (err) {
-          return reject(err?.body ? err.body : err);
-        }
-      }
-
-      const headers = {
-        Accept: 'application/json',
-        'access-token': accessToken,
-      };
-
-      const URL: string = API + `/partners/api/v2/order/${req.body.orderId}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body ? data.body : data);
-          }
-        }
-      );
-    });
-  }
-
-
-  wyreWalletOrderQuotation(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.wyreGetKeys(req);
-      req.body.accountId = keys.ACCOUNT_ID;
-
-      if (req.body.amountIncludeFees) {
-        if (
-          !checkRequired(req.body, ['sourceAmount', 'sourceCurrency', 'destCurrency', 'dest', 'country', 'walletType'])
-        ) {
-          return reject(new ClientError("Wyre's request missing arguments"));
-        }
-      } else {
-        if (!checkRequired(req.body, ['amount', 'sourceCurrency', 'destCurrency', 'dest', 'country'])) {
-          return reject(new ClientError("Wyre's request missing arguments"));
-        }
-      }
-
-      const URL: string = `${keys.API}/v3/orders/quote/partner?timestamp=${Date.now().toString()}`;
-      const XApiSignature: string = URL + JSON.stringify(req.body);
-      const XApiSignatureHash: string = Bitcore.crypto.Hash.sha256hmac(
-        Buffer.from(XApiSignature),
-        Buffer.from(keys.SECRET_API_KEY)
-      ).toString('hex');
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'X-Api-Key': keys.API_KEY,
-        'X-Api-Signature': XApiSignatureHash
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  wyreWalletOrderReservation(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const keys = this.wyreGetKeys(req);
-      req.body.referrerAccountId = keys.ACCOUNT_ID;
-
-      if (req.body.amountIncludeFees) {
-        if (
-          !checkRequired(req.body, [
-            'sourceAmount',
-            'sourceCurrency',
-            'destCurrency',
-            'dest',
-            'country',
-            'paymentMethod'
-          ])
-        ) {
-          return reject(new ClientError("Wyre's request missing arguments"));
-        }
-      } else {
-        if (!checkRequired(req.body, ['amount', 'sourceCurrency', 'destCurrency', 'dest', 'paymentMethod'])) {
-          return reject(new ClientError("Wyre's request missing arguments"));
-        }
-      }
-
-      const URL: string = `${keys.API}/v3/orders/reserve?timestamp=${Date.now().toString()}`;
-      const XApiSignature: string = URL + JSON.stringify(req.body);
-      const XApiSignatureHash: string = Bitcore.crypto.Hash.sha256hmac(
-        Buffer.from(XApiSignature),
-        Buffer.from(keys.SECRET_API_KEY)
-      ).toString('hex');
-
-      const headers = {
-        'Content-Type': 'application/json',
-        'X-Api-Key': keys.API_KEY,
-        'X-Api-Signature': XApiSignatureHash
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: req.body,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ? err.body : err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  private changellyGetKeysV2(req) {
-    if (!config.changelly) {
-      logger.warn('Changelly missing credentials');
-      throw new Error('ClientError: Service not configured.');
-      if (!config.changelly.v2) {
-        logger.warn('Changelly v2 missing credentials');
-        throw new Error('ClientError: Service v2 not configured.');
-      }
-    }
-
-    const keys = {
-      API: config.changelly.v2.api,
-      SECRET: config.changelly.v2.secret
-    };
-
-    return keys;
-  }
-
-  changellySignRequestsV2(message, secret: string) {
-    if (!message || !secret) throw new Error('Missing parameters to sign Changelly v2 request');
-
-    const privateKey = crypto.createPrivateKey({
-      key: Buffer.from(secret, 'hex'),
-      format: 'der',
-      type: 'pkcs8',
-    });
-
-    const publicKey = crypto.createPublicKey(privateKey).export({
-      type: 'pkcs1',
-      format: 'der'
-    });
-
-    const signature = crypto.sign('sha256', Buffer.from(JSON.stringify(message)), privateKey);
-
-    return { signature, publicKey };
-  }
-
-  changellyGetCurrencies(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      let keys, headers;
-      if (req.body.useV2) {
-        keys = this.changellyGetKeysV2(req);
-      } else {
-        return reject(Errors.UPGRADE_NEEDED.withMessage('Credentials expired, please update the app to continue using Changelly services.'));
-      }
-
-      if (!checkRequired(req.body, ['id'])) {
-        return reject(new ClientError('changellyGetCurrencies request missing arguments'));
-      }
-
-      const message = {
-        jsonrpc: '2.0',
-        id: req.body.id,
-        method: req.body.full ? 'getCurrenciesFull' : 'getCurrencies',
-        params: {}
-      };
-
-      const URL: string = keys.API;
-      const { signature, publicKey } = this.changellySignRequestsV2(message, keys.SECRET);
-      headers = {
-        'Content-Type': 'application/json',
-        'X-Api-Key': crypto.createHash('sha256').update(publicKey).digest('base64'),
-        'X-Api-Signature': signature.toString('base64'),
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: message,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ?? err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  changellyGetPairsParams(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      let keys, headers;
-      if (req.body.useV2) {
-        keys = this.changellyGetKeysV2(req);
-      } else {
-        return reject(Errors.UPGRADE_NEEDED.withMessage('Credentials expired, please update the app to continue using Changelly services.'));
-      }
-
-      if (!checkRequired(req.body, ['id', 'coinFrom', 'coinTo'])) {
-        return reject(new ClientError('changellyGetPairsParams request missing arguments'));
-      }
-
-      const message = {
-        id: req.body.id,
-        jsonrpc: '2.0',
-        method: 'getPairsParams',
-        params: [
-          {
-            from: req.body.coinFrom,
-            to: req.body.coinTo
-          }
-        ]
-      };
-
-      const URL: string = keys.API;
-      const { signature, publicKey } = this.changellySignRequestsV2(message, keys.SECRET);
-      headers = {
-        'Content-Type': 'application/json',
-        'X-Api-Key': crypto.createHash('sha256').update(publicKey).digest('base64'),
-        'X-Api-Signature': signature.toString('base64'),
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: message,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ?? err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  changellyGetFixRateForAmount(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      let keys, headers;
-      if (req.body.useV2) {
-        keys = this.changellyGetKeysV2(req);
-      } else {
-        return reject(Errors.UPGRADE_NEEDED.withMessage('Credentials expired, please update the app to continue using Changelly services.'));
-      }
-
-      if (!checkRequired(req.body, ['id', 'coinFrom', 'coinTo', 'amountFrom'])) {
-        return reject(new ClientError('changellyGetFixRateForAmount request missing arguments'));
-      }
-
-      const message = {
-        id: req.body.id,
-        jsonrpc: '2.0',
-        method: 'getFixRateForAmount',
-        params: [
-          {
-            from: req.body.coinFrom,
-            to: req.body.coinTo,
-            amountFrom: req.body.amountFrom
-          }
-        ]
-      };
-
-      const URL: string = keys.API;
-      const { signature, publicKey } = this.changellySignRequestsV2(message, keys.SECRET);
-      headers = {
-        'Content-Type': 'application/json',
-        'X-Api-Key': crypto.createHash('sha256').update(publicKey).digest('base64'),
-        'X-Api-Signature': signature.toString('base64'),
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: message,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ?? err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  changellyCreateFixTransaction(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      let keys, headers;
-      if (req.body.useV2) {
-        keys = this.changellyGetKeysV2(req);
-      } else {
-        return reject(Errors.UPGRADE_NEEDED.withMessage('Credentials expired, please update the app to continue using Changelly services.'));
-      }
-
-      if (
-        !checkRequired(req.body, [
-          'id',
-          'coinFrom',
-          'coinTo',
-          'amountFrom',
-          'addressTo',
-          'fixedRateId',
-          'refundAddress'
-        ])
-      ) {
-        return reject(new ClientError('changellyCreateFixTransaction request missing arguments'));
-      }
-
-      const message = {
-        id: req.body.id,
-        jsonrpc: '2.0',
-        method: 'createFixTransaction',
-        params: {
-          from: req.body.coinFrom,
-          to: req.body.coinTo,
-          address: req.body.addressTo,
-          amountFrom: req.body.amountFrom,
-          rateId: req.body.fixedRateId,
-          refundAddress: req.body.refundAddress
-        }
-      };
-
-      const URL: string = keys.API;
-      const { signature, publicKey } = this.changellySignRequestsV2(message, keys.SECRET);
-      headers = {
-        'Content-Type': 'application/json',
-        'X-Api-Key': crypto.createHash('sha256').update(publicKey).digest('base64'),
-        'X-Api-Signature': signature.toString('base64'),
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: message,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ?? err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  changellyGetTransactions(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      let keys, headers;
-      if (req.body.useV2) {
-        keys = this.changellyGetKeysV2(req);
-      } else {
-        return reject(Errors.UPGRADE_NEEDED.withMessage('Credentials expired, please update the app to continue using Changelly services.'));
-      }
-
-      if (!checkRequired(req.body, ['id', 'exchangeTxId'])) {
-        return reject(new ClientError('changellyGetTransactions request missing arguments'));
-      }
-
-      const message = {
-        id: req.body.id,
-        jsonrpc: '2.0',
-        method: 'getTransactions',
-        params:
-        {
-          id: req.body.exchangeTxId,
-          limit: req.body.limit ?? 1,
-        }
-      };
-
-      const URL: string = keys.API;
-      const { signature, publicKey } = this.changellySignRequestsV2(message, keys.SECRET);
-      headers = {
-        'Content-Type': 'application/json',
-        'X-Api-Key': crypto.createHash('sha256').update(publicKey).digest('base64'),
-        'X-Api-Signature': signature.toString('base64'),
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: message,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ?? err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  changellyGetStatus(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      let keys, headers;
-      if (req.body.useV2) {
-        keys = this.changellyGetKeysV2(req);
-      } else {
-        return reject(Errors.UPGRADE_NEEDED.withMessage('Credentials expired, please update the app to continue using Changelly services.'));
-      }
-
-      if (!checkRequired(req.body, ['id', 'exchangeTxId'])) {
-        return reject(new ClientError('changellyGetStatus request missing arguments'));
-      }
-
-      const message = {
-        jsonrpc: '2.0',
-        id: req.body.id,
-        method: 'getStatus',
-        params: {
-          id: req.body.exchangeTxId
-        }
-      };
-
-      const URL: string = keys.API;
-      const { signature, publicKey } = this.changellySignRequestsV2(message, keys.SECRET);
-      headers = {
-        'Content-Type': 'application/json',
-        'X-Api-Key': crypto.createHash('sha256').update(publicKey).digest('base64'),
-        'X-Api-Signature': signature.toString('base64'),
-      };
-
-      this.request.post(
-        URL,
-        {
-          headers,
-          body: message,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ?? err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  private oneInchGetCredentials() {
-    if (!config.oneInch) throw new Error('1Inch missing credentials');
-
-    const credentials = {
-      API: config.oneInch.api,
-      API_KEY: config.oneInch.apiKey,
-      referrerAddress: config.oneInch.referrerAddress,
-      referrerFee: config.oneInch.referrerFee
-    };
-
-    return credentials;
-  }
-
-  oneInchGetReferrerFee(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const credentials = this.oneInchGetCredentials();
-
-      const referrerFee: number = credentials.referrerFee;
-
-      resolve({ referrerFee });
-    });
-  }
-
-  oneInchGetSwap(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const credentials = this.oneInchGetCredentials();
-
-      if (
-        !checkRequired(req.body, [
-          'fromTokenAddress',
-          'toTokenAddress',
-          'amount',
-          'fromAddress',
-          'slippage',
-          'destReceiver'
-        ])
-      ) {
-        return reject(new ClientError('oneInchGetSwap request missing arguments'));
-      }
-
-      const headers = {
-        'Content-Type': 'application/json'
-      };
-
-      let qs = [];
-      qs.push('fromTokenAddress=' + req.body.fromTokenAddress);
-      qs.push('toTokenAddress=' + req.body.toTokenAddress);
-      qs.push('amount=' + req.body.amount);
-      qs.push('fromAddress=' + req.body.fromAddress);
-      qs.push('slippage=' + req.body.slippage);
-      qs.push('destReceiver=' + req.body.destReceiver);
-
-      if (credentials.referrerFee) qs.push('fee=' + credentials.referrerFee);
-      if (credentials.referrerAddress) qs.push('referrerAddress=' + credentials.referrerAddress);
-
-      const chainNetwork: string = `${req.params?.['chain']?.toUpperCase()}_mainnet` || 'eth_mainnet';
-      const chainId: number = ConstantsCWC.EVM_CHAIN_NETWORK_TO_CHAIN_ID[chainNetwork];
-
-      const URL: string = `${credentials.API}/v5.2/${chainId}/swap/?${qs.join('&')}`;
-
-      this.request.get(
-        URL,
-        {
-          headers,
-          json: true
-        },
-        (err, data) => {
-          if (err) {
-            return reject(err.body ?? err);
-          } else {
-            return resolve(data.body);
-          }
-        }
-      );
-    });
-  }
-
-  oneInchGetTokens(req): Promise<any> {
-    return new Promise((resolve, reject) => {
-
-      const credentials = this.oneInchGetCredentials();
-      const chain = req.params?.['chain'] || 'eth';
-      const cacheKey = `oneInchTokens:${chain}`;
-
-      this.storage.checkAndUseGlobalCache(cacheKey, Defaults.ONE_INCH_CACHE_DURATION, (err, values, oldvalues) => {
-        if (err) this.logw('Could not get stored tokens list', err);
-        if (values) return resolve(values);
-
-        const headers = {
-          'Content-Type': 'application/json',
-          Accept: 'application/json',
-          Authorization: 'Bearer ' + credentials.API_KEY,
-        };
-
-        const chainIdMap = {
-          eth: 1,
-          matic: 137,
-          pol: 137,
-          arb: 42161,
-          base: 8453,
-          op: 10,
-        };
-
-        const chainId = chainIdMap[chain];
-
-        const URL: string = `${credentials.API}/v5.2/${chainId}/tokens`;
-
-        this.request.get(
-          URL,
-          {
-            headers,
-            json: true
-          },
-          (err, data) => {
-            if (err) {
-              this.logw('An error occured while retrieving the token list', err);
-              if (oldvalues) {
-                this.logw('Using old cached values');
-                return resolve(oldvalues);
-              }
-              return reject(err.body ?? err);
-            } else if (data?.statusCode === 429 && oldvalues) {
-              // oneinch rate limit
-              return resolve(oldvalues);
-            } else {
-              if (!data?.body?.tokens) {
-                if (oldvalues) {
-                  this.logw('No token list available... using old cached values');
-                  return resolve(oldvalues);
-                }
-                return reject(new Error('Could not get tokens list'));
-              }
-              this.storage.storeGlobalCache(cacheKey, data.body.tokens, err => {
-                if (err) {
-                  this.logw('Could not store tokens list');
-                }
-                return resolve(data.body.tokens);
-              });
-            }
-          }
-        );
-      });
-    });
-  }
-
   checkServiceAvailability(req): boolean {
     if (!checkRequired(req.body, ['service', 'opts'])) {
       throw new ClientError('checkServiceAvailability request missing arguments');
@@ -7639,7 +5208,7 @@ export class WalletService implements IWalletService {
 }
 
 
-function checkRequired(obj, args, cb?: (e: any) => void) {
+export function checkRequired(obj, args, cb?: (e: any) => void) {
   const missing = Utils.getMissingFields(obj, args);
   if (!missing.length) {
     return true;
