+++ packages/bitcore-wallet-service/test/integration/server.js
@@ -11357,2991 +11357,6 @@ describe('Wallet service', function() {
     });
   });
 
-  describe('Banxa', () => {
-    let server, wallet, fakeRequest, req;
-    beforeEach((done) => {
-      transport.level = 'info';
-
-      config.banxa = {
-        sandbox: {
-          api: 'api1',
-          apiKey: 'apiKey1',
-          secretKey: 'secretKey1',
-        },
-        production: {
-          api: 'api2',
-          apiKey: 'apiKey2',
-          secretKey: 'secretKey2',
-        },
-        sandboxWeb: {
-          api: 'api3',
-          apiKey: 'apiKey3',
-          secretKey: 'secretKey3',
-        },
-        productionWeb: {
-          api: 'api4',
-          apiKey: 'apiKey4',
-          secretKey: 'secretKey4',
-        }
-      }
-
-      fakeRequest = {
-        get: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-        post: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-      };
-
-      helpers.createAndJoinWallet(1, 1, (s, w) => {
-        wallet = w;
-        var priv = TestData.copayers[0].privKey_1H_0;
-        var sig = helpers.signMessage('hello world', priv);
-
-        WalletService.getInstanceWithAuth({
-          // test assumes wallet's copayer[0] is TestData's copayer[0]
-          copayerId: wallet.copayers[0].id,
-          message: 'hello world',
-          signature: sig,
-          clientVersion: 'bwc-2.0.0',
-          walletId: '123',
-        }, (err, s) => {
-          should.not.exist(err);
-          server = s;
-          done();
-        });
-      });
-    });
-
-    describe('#banxaGetPaymentMethods', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetPaymentMethods(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetPaymentMethods(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.banxaGetPaymentMethods(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if banxa is commented in config', async () => {
-        config.banxa = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetPaymentMethods(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Banxa missing credentials');
-        }
-      });
-    });
-
-    describe('#banxaGetQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            source: 'USD',
-            target: 'BTC'
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.banxaGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.target;
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Banxa\'s request missing arguments');
-        }
-      });
-
-      it('should return error if banxa is commented in config', async () => {
-        config.banxa = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Banxa missing credentials');
-        }
-      });
-    });
-
-    describe('#banxaCreateOrder', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            account_reference: 'account_reference1',
-            source: 'USD',
-            target: 'BTC',
-            wallet_address: 'wallet_address1',
-            return_url_on_success: 'return_url_on_success1'
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaCreateOrder(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaCreateOrder(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if post returns error', async () => {
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.banxaCreateOrder(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.source;
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaCreateOrder(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Banxa\'s request missing arguments');
-        }
-      });
-
-      it('should return error if banxa is commented in config', async () => {
-        config.banxa = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaCreateOrder(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Banxa missing credentials');
-        }
-      });
-    });
-
-    describe('#banxaGetOrder', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            order_id: 'order_id1',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetOrder(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetOrder(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.banxaGetOrder(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.order_id;
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetOrder(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Banxa\'s request missing arguments');
-        }
-      });
-
-      it('should return error if banxa is commented in config', async () => {
-        config.banxa = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.banxaGetOrder(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Banxa missing credentials');
-        }
-      });
-    });
-  });
-
-  describe('Moonpay', () => {
-    let server, wallet, fakeRequest, req;
-    beforeEach((done) => {
-      transport.level = 'info';
-
-      config.moonpay = {
-        sandbox: {
-          apiKey: 'apiKey1',
-          api: 'api1',
-          widgetApi: 'widgetApi1',
-          sellWidgetApi: 'sellWidgetApi1',
-          secretKey: 'secretKey1'
-        },
-        production: {
-          apiKey: 'apiKey2',
-          api: 'api2',
-          widgetApi: 'widgetApi2',
-          sellWidgetApi: 'sellWidgetApi2',
-          secretKey: 'secretKey2'
-        },
-        sandboxWeb: {
-          apiKey: 'apiKey3',
-          api: 'api3',
-          widgetApi: 'widgetApi3',
-          sellWidgetApi: 'sellWidgetApi3',
-          secretKey: 'secretKey3'
-        },
-        productionWeb: {
-          apiKey: 'apiKey4',
-          api: 'api4',
-          widgetApi: 'widgetApi4',
-          sellWidgetApi: 'sellWidgetApi4',
-          secretKey: 'secretKey4'
-        }
-      }
-
-      fakeRequest = {
-        get: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-        post: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-        delete: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-      };
-
-      helpers.createAndJoinWallet(1, 1, (s, w) => {
-        wallet = w;
-        var priv = TestData.copayers[0].privKey_1H_0;
-        var sig = helpers.signMessage('hello world', priv);
-
-        WalletService.getInstanceWithAuth({
-          // test assumes wallet's copayer[0] is TestData's copayer[0]
-          copayerId: wallet.copayers[0].id,
-          message: 'hello world',
-          signature: sig,
-          clientVersion: 'bwc-2.0.0',
-          walletId: '123',
-        }, (err, s) => {
-          should.not.exist(err);
-          server = s;
-          done();
-        });
-      });
-    });
-
-    describe('#moonpayGetQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            currencyAbbreviation: 'btc',
-            baseCurrencyAmount: 50,
-            extraFeePercentage: 5,
-            baseCurrencyCode: 'usd'
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.moonpayGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.baseCurrencyAmount;
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay\'s request missing arguments');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', async () => {
-        config.moonpay = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-
-    describe('#moonpayGetSellQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            currencyAbbreviation: 'btc',
-            quoteCurrencyCode: 'usd',
-            baseCurrencyAmount: 1
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetSellQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetSellQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.moonpayGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.baseCurrencyAmount;
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay\'s request missing arguments');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', async () => {
-        config.moonpay = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-
-    describe('#moonpayGetCurrencyLimits', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            currencyAbbreviation: 'btc',
-            baseCurrencyCode: 'usd'
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetCurrencyLimits(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetCurrencyLimits(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.moonpayGetCurrencyLimits(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.baseCurrencyCode;
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetCurrencyLimits(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay\'s request missing arguments');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', async () => {
-        config.moonpay = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetCurrencyLimits(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-
-    describe('#moonpayGetSignedPaymentUrl', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'production',
-            currencyCode: 'btc',
-            walletAddress: 'bitcoin:123123',
-            baseCurrencyCode: 'usd',
-            baseCurrencyAmount: '500',
-            externalTransactionId: '123123',
-            redirectURL: 'bitpay://moonpay'
-          }
-        }
-      });
-
-      it('should get the paymentUrl properly if req is OK', () => {
-        try {
-          const data = server.moonpayGetSignedPaymentUrl(req);
-          should.exist(data.urlWithSignature);
-          data.urlWithSignature.should.equal('widgetApi2?apiKey=apiKey2&currencyCode=btc&walletAddress=bitcoin%3A123123&baseCurrencyCode=usd&baseCurrencyAmount=500&externalTransactionId=123123&redirectURL=bitpay%3A%2F%2Fmoonpay&signature=%2FDnbsboySgE%2FeAvMrwzROCLuuctkhgw5C2t2OofjOzo%3D');
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if there is some missing arguments', () => {
-        delete req.body.currencyCode;
-
-        try {
-          const data = server.moonpayGetSignedPaymentUrl(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay\'s request missing arguments');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', () => {
-        config.moonpay = undefined;
-
-        try {
-          const data = server.moonpayGetSignedPaymentUrl(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-
-    describe('#moonpayGetSellSignedPaymentUrl', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'production',
-            baseCurrencyCode: 'btc',
-            baseCurrencyAmount: 500,
-            externalTransactionId: '123123',
-            redirectURL: 'bitpay://moonpay',
-            quoteCurrencyCode: 'usd',
-            refundWalletAddress: 'bitcoin:123123',
-          }
-        }
-      });
-
-      it('should get the paymentUrl properly if req is OK', () => {
-        try {
-          const data = server.moonpayGetSellSignedPaymentUrl(req);
-          should.exist(data.urlWithSignature);
-          data.urlWithSignature.should.equal('sellWidgetApi2?apiKey=apiKey2&baseCurrencyCode=btc&baseCurrencyAmount=500&externalTransactionId=123123&redirectURL=bitpay%3A%2F%2Fmoonpay&quoteCurrencyCode=usd&refundWalletAddress=bitcoin%3A123123&signature=otiVaKVxKT%2BRNOfkSMOk07U3JxY4DrpPAztiXl5Wvjc%3D');
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if there is some missing arguments', () => {
-        delete req.body.baseCurrencyCode;
-
-        try {
-          const data = server.moonpayGetSellSignedPaymentUrl(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay\'s request missing arguments');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', () => {
-        config.moonpay = undefined;
-
-        try {
-          const data = server.moonpayGetSellSignedPaymentUrl(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-
-    describe('#moonpayGetTransactionDetails', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            transactionId: 'transactionId1',
-          }
-        }
-      });
-
-      it('should work properly if req is OK with transactionId', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetTransactionDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK with externalId', async () => {
-        delete req.body.transactionId;
-        req.body.externalId = 'externalId1';
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetTransactionDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.moonpayGetTransactionDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        }
-      });
-
-      it('should return error if there is no transactionId or externalId', async () => {
-        delete req.body.transactionId;
-        delete req.body.externalId;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetTransactionDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay\'s request missing arguments');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', async () => {
-        config.moonpay = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetTransactionDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-
-    describe('#moonpayGetSellTransactionDetails', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            transactionId: 'transactionId1',
-          }
-        }
-      });
-
-      it('should work properly if req is OK with transactionId', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetSellTransactionDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK with externalId', async () => {
-        delete req.body.transactionId;
-        req.body.externalId = 'externalId1';
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetSellTransactionDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.moonpayGetSellTransactionDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        }
-      });
-
-      it('should return error if there is no transactionId or externalId', async () => {
-        delete req.body.transactionId;
-        delete req.body.externalId;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetSellTransactionDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay\'s request missing arguments');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', async () => {
-        config.moonpay = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetSellTransactionDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-
-    describe('#moonpayGetAccountDetails', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetAccountDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.moonpayGetAccountDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', async () => {
-        config.moonpay = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayGetAccountDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-
-    describe('#moonpayCancelSellTransaction', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            transactionId: 'transactionId1',
-          }
-        }
-      });
-
-      it('should work properly if req is OK with transactionId', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayCancelSellTransaction(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK with externalId', async () => {
-        delete req.body.transactionId;
-        req.body.externalId = 'externalId1';
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayCancelSellTransaction(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if delete returns error', async () => {
-        const fakeRequest2 = {
-          delete: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.moonpayCancelSellTransaction(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        }
-      });
-
-      it('should return error if there is no transactionId or externalId', async () => {
-        delete req.body.transactionId;
-        delete req.body.externalId;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayCancelSellTransaction(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay\'s request missing arguments');
-        }
-      });
-
-      it('should return error if moonpay is commented in config', async () => {
-        config.moonpay = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.moonpayCancelSellTransaction(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Moonpay missing credentials');
-        }
-      });
-    });
-  });
-
-  describe('Sardine', () => {
-    let server, wallet, fakeRequest, req;
-    beforeEach((done) => {
-      transport.level = 'info';
-
-      config.sardine = {
-        sandbox: {
-          api: 'api1',
-          secretKey: 'secretKey1',
-          clientId: 'clientId1',
-        },
-        production: {
-          api: 'api2',
-          secretKey: 'secretKey2',
-          clientId: 'clientId2',
-        },
-        sandboxWeb: {
-          api: 'api3',
-          secretKey: 'secretKey3',
-          clientId: 'clientId3',
-        },
-        productionWeb: {
-          api: 'api4',
-          secretKey: 'secretKey4',
-          clientId: 'clientId4',
-        }
-      }
-
-      fakeRequest = {
-        get: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-        post: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-      };
-
-      helpers.createAndJoinWallet(1, 1, (s, w) => {
-        wallet = w;
-        var priv = TestData.copayers[0].privKey_1H_0;
-        var sig = helpers.signMessage('hello world', priv);
-
-        WalletService.getInstanceWithAuth({
-          // test assumes wallet's copayer[0] is TestData's copayer[0]
-          copayerId: wallet.copayers[0].id,
-          message: 'hello world',
-          signature: sig,
-          clientVersion: 'bwc-2.0.0',
-          walletId: '123',
-        }, (err, s) => {
-          should.not.exist(err);
-          server = s;
-          done();
-        });
-      });
-    });
-
-    describe('#sardineGetQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            asset_type: 'BTC',
-            network: 'bitcoin',
-            total: 50,
-            currency: 'USD',
-            paymentType: 'debit',
-            quote_type: 'buy'
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.sardineGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.asset_type;
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Sardine\'s request missing arguments');
-        }
-      });
-
-      it('should return error if sardine is commented in config', async () => {
-        config.sardine = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Sardine missing credentials');
-        }
-      });
-    });
-
-    describe('#sardineGetCurrencyLimits', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetCurrencyLimits(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetCurrencyLimits(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.sardineGetCurrencyLimits(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if sardine is commented in config', async () => {
-        config.sardine = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetCurrencyLimits(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Sardine missing credentials');
-        }
-      });
-    });
-
-    describe('#sardineGetToken', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            referenceId: 'referenceId1',
-            externalUserId: 'externalUserId1',
-            customerId: 'customerId1',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetToken(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetToken(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if post returns error', async () => {
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.sardineGetToken(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.referenceId;
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetToken(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Sardine\'s request missing arguments');
-        }
-      });
-
-      it('should return error if sardine is commented in config', async () => {
-        config.sardine = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetToken(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Sardine missing credentials');
-        }
-      });
-    });
-
-    describe('#sardineGetSupportedTokens', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async() => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetSupportedTokens(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async() => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetSupportedTokens(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async() => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.sardineGetSupportedTokens(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if sardine is commented in config', async() => {
-        config.sardine = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetSupportedTokens(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Sardine missing credentials');
-        }
-      });
-    });
-
-    describe('#sardineGetOrdersDetails', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            orderId: 'orderId1',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetOrdersDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetOrdersDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.sardineGetOrdersDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.orderId;
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetOrdersDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Sardine\'s request missing arguments');
-        }
-      });
-
-      it('should work properly if orderId is not present but externalUserId is', async () => {
-        delete req.body.orderId;
-        req.body.externalUserId = 'externalUserId1';
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetOrdersDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if sardine is commented in config', async () => {
-        config.sardine = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.sardineGetOrdersDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Sardine missing credentials');
-        }
-      });
-    });
-  });
-
-  describe('Ramp', () => {
-    let server, wallet, fakeRequest, req;
-    beforeEach((done) => {
-      transport.level = 'info';
-
-      config.ramp = {
-        sandbox: {
-          apiKey: 'apiKey1',
-          api: 'api1',
-          widgetApi: 'widgetApi1',
-        },
-        production: {
-          apiKey: 'apiKey2',
-          api: 'api2',
-          widgetApi: 'widgetApi2',
-        },
-        sandboxWeb: {
-          apiKey: 'apiKey3',
-          api: 'api3',
-          widgetApi: 'widgetApi3',
-        },
-        productionWeb: {
-          apiKey: 'apiKey4',
-          api: 'api4',
-          widgetApi: 'widgetApi4',
-        }
-      }
-
-      fakeRequest = {
-        get: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-        post: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-      };
-
-      helpers.createAndJoinWallet(1, 1, (s, w) => {
-        wallet = w;
-        var priv = TestData.copayers[0].privKey_1H_0;
-        var sig = helpers.signMessage('hello world', priv);
-
-        WalletService.getInstanceWithAuth({
-          // test assumes wallet's copayer[0] is TestData's copayer[0]
-          copayerId: wallet.copayers[0].id,
-          message: 'hello world',
-          signature: sig,
-          clientVersion: 'bwc-2.0.0',
-          walletId: '123',
-        }, (err, s) => {
-          should.not.exist(err);
-          server = s;
-          done();
-        });
-      });
-    });
-
-    describe('#rampGetQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            cryptoAssetSymbol: 'BTC_BTC',
-            fiatValue: 50,
-            fiatCurrency: 'USD',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if post returns error', async () => {
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.rampGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.fiatValue;
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Ramp\'s request missing arguments');
-        }
-      });
-
-      it('should return error if ramp is commented in config', async () => {
-        config.ramp = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Ramp missing credentials');
-        }
-      });
-    });
-
-    describe('#rampGetSellQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            cryptoAssetSymbol: 'BTC_BTC',
-            cryptoAmount: '10000000',
-            fiatCurrency: 'USD',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetSellQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetSellQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if post returns error', async () => {
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.rampGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.cryptoAmount;
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Ramp\'s request missing arguments');
-        }
-      });
-
-      it('should return error if ramp is commented in config', async () => {
-        config.ramp = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Ramp missing credentials');
-        }
-      });
-    });
-
-    describe('#rampGetSignedPaymentUrl', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'production',
-            flow: 'buy',
-            swapAsset: 'BTC_BTC',
-            swapAmount: '1000000',
-            enabledFlows: 'ONRAMP',
-            defaultFlow: 'ONRAMP',
-            userAddress: 'bitcoin:123123',
-            selectedCountryCode: 'US',
-            defaultAsset: 'BTC_BTC',
-            finalUrl: 'bitpay://ramp',
-          }
-        }
-      });
-
-      it('should get the paymentUrl properly if req is OK', () => {
-        try {
-          const data = server.rampGetSignedPaymentUrl(req);
-          should.exist(data.urlWithSignature);
-          data.urlWithSignature.should.equal('widgetApi2?hostApiKey=apiKey2&selectedCountryCode=US&finalUrl=bitpay%3A%2F%2Framp&userAddress=bitcoin%3A123123&swapAsset=BTC_BTC&enabledFlows=ONRAMP&defaultFlow=ONRAMP&swapAmount=1000000&defaultAsset=BTC_BTC');
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should get the paymentUrl properly if req is OK for web', () => {
-        try {
-          req.body = {
-            env: 'production',
-            context: 'web',
-            swapAsset: 'BTC_BTC',
-            userAddress: 'bitcoin:123123',
-            selectedCountryCode: 'US',
-            defaultAsset: 'BTC_BTC',
-            finalUrl: 'bitpay://ramp',
-          }
-          const data = server.rampGetSignedPaymentUrl(req);
-          should.exist(data.urlWithSignature);
-          data.urlWithSignature.should.equal('widgetApi4?hostApiKey=apiKey4&selectedCountryCode=US&finalUrl=bitpay%3A%2F%2Framp&userAddress=bitcoin%3A123123&swapAsset=BTC_BTC&defaultAsset=BTC_BTC');
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if there is some missing arguments', () => {
-        delete req.body.defaultAsset;
-
-        try {
-          const data = server.rampGetSignedPaymentUrl(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Ramp\'s request missing arguments');
-        }
-      });
-
-      it('should return error if ramp is commented in config', () => {
-        config.ramp = undefined;
-
-        try {
-          const data = server.rampGetSignedPaymentUrl(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Ramp missing credentials');
-        }
-      });
-
-      it('should get the sell paymentUrl properly if req is OK', () => {
-        try {
-          req.body = {
-            env: 'production',
-            flow: 'sell',
-            offrampAsset: 'BTC_BTC',
-            swapAmount: '1000000',
-            enabledFlows: 'OFFRAMP',
-            defaultFlow: 'OFFRAMP',
-            selectedCountryCode: 'US',
-            defaultAsset: 'BTC_BTC',
-            variant: 'webview-mobile',
-            useSendCryptoCallback: true,
-            useSendCryptoCallbackVersion: 1,
-            hideExitButton: false,
-          }
-          const data = server.rampGetSignedPaymentUrl(req);
-          should.exist(data.urlWithSignature);
-          data.urlWithSignature.should.equal('widgetApi2?hostApiKey=apiKey2&selectedCountryCode=US&offrampAsset=BTC_BTC&enabledFlows=OFFRAMP&defaultFlow=OFFRAMP&swapAmount=1000000&defaultAsset=BTC_BTC&useSendCryptoCallback=true&variant=webview-mobile&useSendCryptoCallbackVersion=1');
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should get the sell paymentUrl properly if req is OK for web', () => {
-        try {
-          req.body = {
-            env: 'production',
-            flow: 'sell',
-            context: 'web',
-            offrampAsset: 'BTC_BTC',
-            swapAmount: '1000000',
-            enabledFlows: 'OFFRAMP',
-            defaultFlow: 'OFFRAMP',
-            selectedCountryCode: 'US',
-            defaultAsset: 'BTC_BTC',
-            variant: 'webview-mobile',
-            useSendCryptoCallback: true,
-            useSendCryptoCallbackVersion: 1,
-            hideExitButton: false,
-          }
-          const data = server.rampGetSignedPaymentUrl(req);
-          should.exist(data.urlWithSignature);
-          data.urlWithSignature.should.equal('widgetApi4?hostApiKey=apiKey4&selectedCountryCode=US&offrampAsset=BTC_BTC&enabledFlows=OFFRAMP&defaultFlow=OFFRAMP&swapAmount=1000000&defaultAsset=BTC_BTC&useSendCryptoCallback=true&variant=webview-mobile&useSendCryptoCallbackVersion=1');
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-    });
-
-    describe('#rampGetAssets', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            currencyCode: 'USD',
-          }
-        }
-      });
-
-      it('should work properly if req is OK with currencyCode', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetAssets(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK with useIp', async () => {
-        delete req.body.currencyCode;
-        req.body.useIp = true;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetAssets(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.rampGetAssets(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        }
-      });
-
-      it('should return error if Ramp is commented in config', async () => {
-        config.ramp = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetAssets(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Ramp missing credentials');
-        }
-      });
-    });
-
-    describe('#rampGetSellTransactionDetails', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'production',
-            id: 'id1',
-            saleViewToken: 'saleViewToken1',
-          }
-        }
-      });
-
-      it('should work properly if req is OK with currencyCode', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetSellTransactionDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.rampGetSellTransactionDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        }
-      });
-
-      it('should return error if Ramp is commented in config', async () => {
-        config.ramp = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.rampGetSellTransactionDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Ramp missing credentials');
-        }
-      });
-    });
-  });
-
-  describe('Simplex', () => {
-    let server, wallet, fakeRequest, req;
-    beforeEach((done) => {
-      transport.level = 'info';
-
-      config.simplex = {
-        sandbox: {
-          apiKey: 'xxxx',
-          api: 'xxxx',
-          appProviderId: 'xxxx'
-        },
-        production: {
-          apiKey: 'xxxx',
-          api: 'xxxx',
-          appProviderId: 'xxxx'
-        },
-        sandboxWeb: {
-          apiKey: 'xxxx',
-          api: 'xxxx',
-          appProviderId: 'xxxx'
-        },
-        productionWeb: {
-          apiKey: 'xxxx',
-          api: 'xxxx',
-          appProviderId: 'xxxx'
-        }
-      }
-
-      fakeRequest = {
-        get: (_url, _opts, _cb) => { return _cb(null, { data: 'data' }) },
-        post: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-      };
-
-      helpers.createAndJoinWallet(1, 1, (s, w) => {
-        wallet = w;
-        var priv = TestData.copayers[0].privKey_1H_0;
-        var sig = helpers.signMessage('hello world', priv);
-
-        WalletService.getInstanceWithAuth({
-          // test assumes wallet's copayer[0] is TestData's copayer[0]
-          copayerId: wallet.copayers[0].id,
-          message: 'hello world',
-          signature: sig,
-          clientVersion: 'bwc-2.0.0',
-          walletId: '123',
-        }, (err, s) => {
-          should.not.exist(err);
-          server = s;
-          done();
-        });
-      });
-    });
-
-    describe('#simplexGetCurrencies', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox'
-          },
-        }
-      });
-
-      it('should work properly if req is OK', () => {
-        server.request = fakeRequest;
-        server.simplexGetCurrencies(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-
-      it('should work properly if req is OK for web', () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        server.simplexGetCurrencies(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-
-      it('should return error if get returns error', () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        server.simplexGetCurrencies(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Error');
-        });
-      });
-
-      it('should return error if simplex is commented in config', () => {
-        config.simplex = undefined;
-
-        server.request = fakeRequest;
-        server.simplexGetCurrencies(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Simplex missing credentials');
-        });
-      });
-    });
-
-    describe('#simplexGetQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox'
-          },
-          ip: '1.2.3.4'
-        }
-      });
-
-      it('should work properly if req is OK', () => {
-        server.request = fakeRequest;
-        server.simplexGetQuote(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-
-      it('should work properly if req is OK for web', () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        server.simplexGetQuote(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-
-      it('should return error if post returns error', () => {
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        server.simplexGetQuote(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Error');
-        });
-      });
-
-      it('should return error if simplex is commented in config', () => {
-        config.simplex = undefined;
-
-        server.request = fakeRequest;
-        server.simplexGetQuote(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Simplex missing credentials');
-        });
-      });
-    });
-
-    describe('#simplexGetSellQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            userCountry: 'LT',
-            base_currency: 'BTC',
-            base_amount: 1000000,
-            quote_currency: 'EUR',
-            pp_payment_method: 'sepa'
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.simplexGetSellQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.simplexGetSellQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.simplexGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.base_amount;
-        server.request = fakeRequest;
-        try {
-          const data = await server.simplexGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Simplex\'s request missing arguments');
-        }
-      });
-
-      it('should return error if simplex is commented in config', async () => {
-        config.simplex = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.simplexGetSellQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Simplex missing credentials');
-        }
-      });
-    });
-
-    describe('#simplexPaymentRequest', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'production',
-            account_details: {
-            },
-            transaction_details: {
-              payment_details: {
-              }
-            }
-          },
-          ip: '1.2.3.4'
-        }
-
-        fakeRequest = {
-          post: (_url, _opts, _cb) => { return _cb(null, { body: {} }) },
-        };
-      });
-
-      it('should work properly if req is OK', () => {
-        server.request = fakeRequest;
-        server.simplexPaymentRequest(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-
-      it('should return error if post returns error', () => {
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        server.simplexPaymentRequest(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Error');
-        });
-      });
-
-      it('should return error if there is some missing arguments', () => {
-        delete req.body.transaction_details;
-
-        server.request = fakeRequest;
-        server.simplexPaymentRequest(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Simplex\'s request missing arguments');
-        });
-      });
-
-      it('should return error if simplex is commented in config', () => {
-        config.simplex = undefined;
-
-        server.request = fakeRequest;
-        server.simplexPaymentRequest(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Simplex missing credentials');
-        });
-      });
-    });
-
-    describe('#simplexSellPaymentRequest', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'production',     
-            userCountry: 'LT',
-            referer_url: 'https://referer_url.com/',
-            return_url: 'https://return_url.com/',
-            txn_details: {quote_id: 'quote_id_1'},
-          },
-          ip: '1.2.3.4'
-        }
-
-        fakeRequest = {
-          post: (_url, _opts, _cb) => { return _cb(null, { body: {} }) },
-        };
-      });
-
-      it('should work properly if req is OK', () => {
-        server.request = fakeRequest;
-        server.simplexSellPaymentRequest(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-
-      it('should return error if post returns error', () => {
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        server.simplexSellPaymentRequest(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Error');
-        });
-      });
-
-      it('should return error if there is some missing arguments', () => {
-        delete req.body.return_url;
-
-        server.request = fakeRequest;
-        server.simplexSellPaymentRequest(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Simplex\'s request missing arguments');
-        });
-      });
-
-      it('should return error if simplex is commented in config', () => {
-        config.simplex = undefined;
-
-        server.request = fakeRequest;
-        server.simplexSellPaymentRequest(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Simplex missing credentials');
-        });
-      });
-    });
-
-    describe('#simplexGetEvents', () => {
-      beforeEach(() => {
-        req = {
-          env: 'production'
-        }
-
-        fakeRequest = {
-          get: (_url, _opts, _cb) => { return _cb(null, { body: {} }) },
-        };
-      });
-
-      it('should work properly if req is OK', () => {
-        server.request = fakeRequest;
-        server.simplexGetEvents(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-    });
-
-  });
-
-  describe('Transak', () => {
-    let server, wallet, fakeRequest, req;
-    beforeEach((done) => {
-      transport.level = 'info';
-
-      config.transak = {
-        sandbox: {
-          api: 'api1',
-          apiKey: 'apiKey1',
-          secretKey: 'secretKey1',
-          widgetApi: 'widgetApi1',
-        },
-        production: {
-          api: 'api2',
-          apiKey: 'apiKey2',
-          secretKey: 'secretKey2',
-          widgetApi: 'widgetApi2',
-        },
-        sandboxWeb: {
-          api: 'api3',
-          apiKey: 'apiKey3',
-          secretKey: 'secretKey3',
-          widgetApi: 'widgetApi3',
-        },
-        productionWeb: {
-          api: 'api4',
-          apiKey: 'apiKey4',
-          secretKey: 'secretKey4',
-          widgetApi: 'widgetApi4',
-        }
-      }
-
-      fakeRequest = {
-        get: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-        post: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-      };
-
-      helpers.createAndJoinWallet(1, 1, (s, w) => {
-        wallet = w;
-        var priv = TestData.copayers[0].privKey_1H_0;
-        var sig = helpers.signMessage('hello world', priv);
-
-        WalletService.getInstanceWithAuth({
-          // test assumes wallet's copayer[0] is TestData's copayer[0]
-          copayerId: wallet.copayers[0].id,
-          message: 'hello world',
-          signature: sig,
-          clientVersion: 'bwc-2.0.0',
-          walletId: '123',
-        }, (err, s) => {
-          should.not.exist(err);
-          server = s;
-          done();
-        });
-      });
-    });
-
-    describe('#transakGetAccessToken', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetAccessToken(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetAccessToken(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if post returns error', async () => {
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.transakGetAccessToken(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if transak is commented in config', async () => {
-        config.transak = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetAccessToken(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Transak missing credentials');
-        }
-      });
-    });
-
-    describe('#transakGetCryptoCurrencies', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetCryptoCurrencies(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetCryptoCurrencies(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.transakGetCryptoCurrencies(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if transak is commented in config', async () => {
-        config.transak = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetCryptoCurrencies(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Transak missing credentials');
-        }
-      });
-    });
-
-    describe('#transakGetFiatCurrencies', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetFiatCurrencies(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetFiatCurrencies(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.transakGetFiatCurrencies(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if transak is commented in config', async () => {
-        config.transak = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetFiatCurrencies(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Transak missing credentials');
-        }
-      });
-    });
-
-    describe('#transakGetQuote', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            fiatCurrency: 'USD',
-            cryptoCurrency: 'BTC',
-            network: 'mainnet',
-            paymentMethod: 'credit_debit_card'
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetQuote(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.transakGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.fiatCurrency;
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Transak\'s request missing arguments');
-        }
-      });
-
-      it('should return error if transak is commented in config', async () => {
-        config.transak = undefined;
-
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetQuote(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Transak missing credentials');
-        }
-      });
-    });
-
-    describe('#transakGetSignedPaymentUrl', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'production',
-            walletAddress: 'walletAddress1',
-            redirectURL: 'bitpay://transak',
-            fiatAmount: '500',
-            fiatCurrency: 'USD',
-            network: 'mainnet',
-            cryptoCurrencyCode: 'BTC',
-            partnerOrderId: 'partnerOrderId1',
-            partnerCustomerId: 'partnerCustomerId1',
-          }
-        }
-      });
-
-      it('should get the paymentUrl properly if req is OK', () => {
-        try {
-          const data = server.transakGetSignedPaymentUrl(req);
-          should.exist(data.urlWithSignature);
-          data.urlWithSignature.should.equal('widgetApi2?apiKey=apiKey2&walletAddress=walletAddress1&redirectURL=bitpay%3A%2F%2Ftransak&fiatAmount=500&fiatCurrency=USD&network=mainnet&cryptoCurrencyCode=BTC&partnerOrderId=partnerOrderId1&partnerCustomerId=partnerCustomerId1');
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should get the paymentUrl properly if req is OK for web', () => {
-        try {
-          req.body = {
-            env: 'production',
-            context: 'web',
-            walletAddress: 'walletAddress1',
-            redirectURL: 'bitpay://transak',
-            fiatAmount: '500',
-            fiatCurrency: 'USD',
-            network: 'mainnet',
-            cryptoCurrencyCode: 'BTC',
-            partnerOrderId: 'partnerOrderId1',
-            partnerCustomerId: 'partnerCustomerId1',
-          }
-          const data = server.transakGetSignedPaymentUrl(req);
-          should.exist(data.urlWithSignature);
-          data.urlWithSignature.should.equal('widgetApi4?apiKey=apiKey4&walletAddress=walletAddress1&redirectURL=bitpay%3A%2F%2Ftransak&fiatAmount=500&fiatCurrency=USD&network=mainnet&cryptoCurrencyCode=BTC&partnerOrderId=partnerOrderId1&partnerCustomerId=partnerCustomerId1');
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if there is some missing arguments', () => {
-        delete req.body.context;
-        delete req.body.fiatAmount;
-
-        try {
-          const data = server.transakGetSignedPaymentUrl(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Transak\'s request missing arguments');
-        }
-      });
-
-      it('should return error if transak is commented in config', () => {
-        config.transak = undefined;
-
-        try {
-          const data = server.transakGetSignedPaymentUrl(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Transak missing credentials');
-        }
-      });
-    });
-
-    describe('#transakGetOrderDetails', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            orderId: 'orderId1',
-            accessToken: 'accessToken1',
-          }
-        }
-      });
-
-      it('should work properly if req is OK', async () => {
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetOrderDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should work properly if req is OK for web', async () => {
-        req.body.context = 'web';
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetOrderDetails(req);
-          should.exist(data);
-        } catch (err) {
-          should.not.exist(err);
-        }
-      });
-
-      it('should return error if get returns error', async () => {
-        const fakeRequest2 = {
-          get: (_url, _opts, _cb) => { return _cb(new Error('Error'), null) },
-        };
-
-        server.request = fakeRequest2;
-        try {
-          const data = await server.transakGetOrderDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Error');
-        };
-      });
-
-      it('should return error if there is some missing arguments', async () => {
-        delete req.body.orderId;
-        server.request = fakeRequest;
-        try {
-          const data = await server.transakGetOrderDetails(req);
-          should.not.exist(data);
-        } catch (err) {
-          should.exist(err);
-          err.message.should.equal('Transak\'s request missing arguments');
-        }
-      });
-    });
-  });
-
-  describe('Wyre', () => {
-    let server, wallet, fakeRequest, req;
-    beforeEach((done) => {
-      config.wyre = {
-        sandbox: {
-          apiKey: 'xxxx',
-          secretApiKey: 'xxxx',
-          api: 'xxxx',
-          widgetUrl: 'xxxx',
-          appProviderAccountId: 'xxxx'
-        },
-        production: {
-          apiKey: 'xxxx',
-          secretApiKey: 'xxxx',
-          api: 'xxxx',
-          widgetUrl: 'xxxx',
-          appProviderAccountId: 'xxxx'
-        }
-      }
-
-      fakeRequest = {
-        get: (_url, _opts, _cb) => { return _cb(null, { data: 'data' }) },
-        post: (_url, _opts, _cb) => { return _cb(null, { body: 'data' }) },
-      };
-
-      helpers.createAndJoinWallet(1, 1, (s, w) => {
-        wallet = w;
-        var priv = TestData.copayers[0].privKey_1H_0;
-        var sig = helpers.signMessage('hello world', priv);
-
-        WalletService.getInstanceWithAuth({
-          // test assumes wallet's copayer[0] is TestData's copayer[0]
-          copayerId: wallet.copayers[0].id,
-          message: 'hello world',
-          signature: sig,
-          clientVersion: 'bwc-2.0.0',
-          walletId: '123',
-        }, (err, s) => {
-          should.not.exist(err);
-          server = s;
-          done();
-        });
-      });
-    });
-
-    describe('#wyreWalletOrderQuotation', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            amount: 50,
-            sourceCurrency: 'USD',
-            destCurrency: 'BTC',
-            dest: 'bitcoin:123123123',
-            country: 'US'
-          }
-        }
-      });
-
-      it('should work properly if req is OK', () => {
-        server.request = fakeRequest;
-        server.wyreWalletOrderQuotation(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-
-      it('should return error if there is some missing arguments', () => {
-        delete req.body.amount;
-
-        server.request = fakeRequest;
-        server.wyreWalletOrderQuotation(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Wyre\'s request missing arguments');
-        });
-      });
-
-      it('should return error if post returns error', () => {
-        req.body.amount = 50;
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error')) },
-        };
-
-        server.request = fakeRequest2;
-        server.wyreWalletOrderQuotation(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Error');
-        });
-      });
-
-      it('should return error if Wyre is commented in config', () => {
-        config.wyre = undefined;
-
-        server.request = fakeRequest;
-        server.wyreWalletOrderQuotation(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Wyre missing credentials');
-        });
-      });
-
-      it('should return error if amountIncludeFees is true but sourceAmount is not present', () => {
-        req = {
-          headers: {},
-          body: {
-            amountIncludeFees: true,
-            env: 'sandbox',
-            amount: 50,
-            sourceCurrency: 'USD',
-            destCurrency: 'BTC',
-            dest: 'bitcoin:123123123',
-            country: 'US',
-            walletType: 'DEBIT_CARD'
-          }
-        }
-
-        server.request = fakeRequest;
-        server.wyreWalletOrderQuotation(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Wyre\'s request missing arguments');
-        });
-      });
-
-      it('should work properly if req is OK with amountIncludeFees and sourceAmount', () => {
-        req = {
-          headers: {},
-          body: {
-            amountIncludeFees: true,
-            env: 'sandbox',
-            sourceAmount: 50,
-            sourceCurrency: 'USD',
-            destCurrency: 'BTC',
-            dest: 'bitcoin:123123123',
-            country: 'US',
-            walletType: 'DEBIT_CARD'
-          }
-        }
-
-        server.request = fakeRequest;
-        server.wyreWalletOrderQuotation(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-    });
-
-    describe('#wyreWalletOrderReservation', () => {
-      beforeEach(() => {
-        req = {
-          headers: {},
-          body: {
-            env: 'sandbox',
-            amount: 50,
-            sourceCurrency: 'USD',
-            destCurrency: 'BTC',
-            dest: 'bitcoin:123123123',
-            paymentMethod: 'debit-card'
-          }
-        }
-
-        fakeRequest = {
-          post: (_url, _opts, _cb) => { return _cb(null, { body: {} }) },
-        };
-      });
-
-      it('should work properly if req is OK', () => {
-        server.request = fakeRequest;
-        server.wyreWalletOrderReservation(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-
-      it('should return error if there is some missing arguments', () => {
-        delete req.body.amount;
-
-        server.request = fakeRequest;
-        server.wyreWalletOrderReservation(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Wyre\'s request missing arguments');
-        });
-      });
-
-      it('should return error if post returns error', () => {
-        req.body.amount = 50;
-        const fakeRequest2 = {
-          post: (_url, _opts, _cb) => { return _cb(new Error('Error')) },
-        };
-
-        server.request = fakeRequest2;
-        server.wyreWalletOrderReservation(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Error');
-        });
-      });
-
-      it('should return error if Wyre is commented in config', () => {
-        config.wyre = undefined;
-
-        server.request = fakeRequest;
-        server.wyreWalletOrderReservation(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Wyre missing credentials');
-        });
-      });
-
-      it('should return error if amountIncludeFees is true but sourceAmount is not present', () => {
-        req = {
-          headers: {},
-          body: {
-            amountIncludeFees: true,
-            env: 'sandbox',
-            amount: 50,
-            sourceCurrency: 'USD',
-            destCurrency: 'BTC',
-            dest: 'bitcoin:123123123',
-            country: 'US',
-            paymentMethod: 'debit-card'
-          }
-        }
-
-        server.request = fakeRequest;
-        server.wyreWalletOrderReservation(req).then(data => {
-          should.not.exist(data);
-        }).catch(err => {
-          should.exist(err);
-          err.message.should.equal('Wyre\'s request missing arguments');
-        });
-      });
-
-      it('should work properly if req is OK with amountIncludeFees and sourceAmount', () => {
-        req = {
-          headers: {},
-          body: {
-            amountIncludeFees: true,
-            env: 'sandbox',
-            sourceAmount: 50,
-            sourceCurrency: 'USD',
-            destCurrency: 'BTC',
-            dest: 'bitcoin:123123123',
-            country: 'US',
-            paymentMethod: 'debit-card'
-          }
-        }
-
-        server.request = fakeRequest;
-        server.wyreWalletOrderReservation(req).then(data => {
-          should.exist(data);
-        }).catch(err => {
-          should.not.exist(err);
-        });
-      });
-    });
-  });
-
   describe('#getCoinsForTx', function() {
     let server, wallet;
     beforeEach(function() {
